% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accessibility_helpers.R
\name{add_ggplot_alt_caption}
\alias{add_ggplot_alt_caption}
\title{Add alt text as ggplot caption}
\usage{
add_ggplot_alt_caption(plot_object, alt_text, width = 180)
}
\arguments{
\item{plot_object}{ggplot object.}

\item{alt_text}{Character. Alt text.}

\item{width}{Integer. Wrap width for alt text caption. Defaults to 180}
}
\value{
ggplot object.
}
\description{
Add alt text as ggplot caption
}
\keyword{internal}
