% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_utils.R
\name{aggregate_by_column}
\alias{aggregate_by_column}
\title{Aggregate a dataframe into multiple dataframes, based on a columns value.}
\usage{
aggregate_by_column(df, column_name)
}
\arguments{
\item{df}{The dataframe to aggregate.}

\item{column_name}{The column to aggregate the data by.}
}
\value{
A list of dataframes, split up based on the value of column_name.
}
\description{
Aggregate a dataframe into multiple dataframes, based on a columns value.
}
