% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{air_pollution_descriptive_stats}
\alias{air_pollution_descriptive_stats}
\title{Show descriptive statistics}
\usage{
air_pollution_descriptive_stats(
  data,
  variables,
  bin_width = 1,
  output_dir = NULL,
  save_outputs = FALSE
)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate, environmental
and health data}

\item{variables}{Character or character vector with variable(s) to produce
summary statistics for. Must include at least 1 variable.}

\item{bin_width}{Integer. Width of each bin in a histogram of the outcome
variable. Defaults to 1.}

\item{output_dir}{Character. Directory to save descriptive statistics.
Defaults to NULL.}

\item{save_outputs}{Logical. Whether to save outputs. Defaults to FALSE.}
}
\description{
Generates summary statistics for climate, environmental and health data.
}
