% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{air_pollution_power_list}
\alias{air_pollution_power_list}
\title{Air Pollution Power Calculation using Meta Results}
\usage{
air_pollution_power_list(
  meta_results,
  data_with_lags,
  ref_pm25 = 15,
  attr_thr = 95,
  include_national = TRUE
)
}
\arguments{
\item{meta_results}{Meta-analysis results from air_pollution_meta_analysis}

\item{data_with_lags}{Lagged data frame}

\item{ref_pm25}{Numeric. Reference PM2.5 value for attributable risk calculation}

\item{attr_thr}{Integer. Percentile at which to define the PM2.5 threshold for
calculating attributable risk. Defaults to 95.}

\item{include_national}{Logical. Whether to include national level calculations.
Defaults to TRUE.}
}
\value{
A list containing power information by region
}
\description{
Produce a power statistic by region for PM2.5 attributable mortality
using meta-analysis results
}
\keyword{internal}
