% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrdl.R
\name{an_attrdl}
\alias{an_attrdl}
\title{FUNCTION FOR COMPUTING ATTRIBUTABLE MEASURES FROM DLNM}
\usage{
an_attrdl(
  x,
  basis,
  cases,
  coef = NULL,
  vcov = NULL,
  model.link = NULL,
  dir = "back",
  tot = TRUE,
  cen,
  range = NULL,
  nsim = 5000
)
}
\arguments{
\item{x}{AN EXPOSURE VECTOR OR (ONLY FOR dir="back") A MATRIX OF LAGGED EXPOSURES}

\item{basis}{THE CROSS-BASIS COMPUTED FROM x}

\item{cases}{THE CASES VECTOR OR (ONLY FOR dir="forw") THE MATRIX OF FUTURE CASES}

\item{coef}{COEF FOR basis IF model IS NOT PROVIDED}

\item{vcov}{VCOV FOR basis IF model IS NOT PROVIDED}

\item{model.link}{LINK FUNCTION IF model IS NOT PROVIDED}

\item{dir}{EITHER "back" OR "forw" FOR BACKWARD OR FORWARD PERSPECTIVES}

\item{tot}{IF TRUE, THE TOTAL ATTRIBUTABLE RISK IS COMPUTED}

\item{cen}{THE REFERENCE VALUE USED AS COUNTERFACTUAL SCENARIO}

\item{range}{THE RANGE OF EXPOSURE. IF NULL, THE WHOLE RANGE IS USED}

\item{nsim}{NUMBER OF SIMULATION SAMPLES}
}
\value{
\itemize{
\item Attributable Fraction
\item Attributable Fraction lower confidence intervals
\item Attributable Fraction upper confidence intervals
\item Attributable Numbers
\item Attributable Numbers lower confidence intervals
\item Attributable Numbers upper confidence intervals
\item Simulation matrix of attributable numbers
}
}
\description{
A function to calculate attributable numbers and fractions derived from
(c) Antonio Gasparrini 2015-2017. Modifications to produce daily values with
confidence intervals.
}
\keyword{internal}
