% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{analyze_air_pollution_daily}
\alias{analyze_air_pollution_daily}
\title{Calculate daily RR/AF/AN/AR for region-specific/national distributed lag effects
for a chosen PM2.5 reference.}
\usage{
analyze_air_pollution_daily(
  data_with_lags,
  meta_results,
  ref_pm25 = 15,
  ref_name = "WHO",
  max_lag = 14L
)
}
\arguments{
\item{data_with_lags}{Dataset. Lagged data with lag variables.}

\item{meta_results}{Dataset. Results from meta analysis.}

\item{ref_pm25}{Numeric. PM2.5 reference value. Defaults to 15.}

\item{ref_name}{Character. Reference body name. Defaults to "WHO".}

\item{max_lag}{Integer. Maximum lag days. Defaults to 14.}
}
\value{
List with region-specific/national results for daily RR/AF/AN/AR
}
\description{
Calculate daily RR/AF/AN/AR for region-specific/national distributed lag effects
for a chosen PM2.5 reference.
}
\keyword{internal}
