% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{analyze_air_pollution_lags}
\alias{analyze_air_pollution_lags}
\title{Analyze multiple lag structures with AF and AN}
\usage{
analyze_air_pollution_lags(data, max_lag = 2, family = "quasipoisson")
}
\arguments{
\item{data}{Input data}

\item{max_lag}{Integer. Maximum lag days. Defaults to 2}

\item{family}{Character string indicating the distribution family used in the GAM.
Must be one of the four options "quasipoisson", "nb", "ziP", and "poisson".
"quasipoisson" (Appropriate for count data with overdispersion),
"nb" (negativebinomial which is suitable for highly dispersed count data),
"ziP" (zero-inflated which is suitable for count data with excess zeros (sparse data)),
"poisson" (Suitable for count data without overdispersion (mean = variance)).
Defaults to "quasipoisson"}
}
\value{
Dataframe with lag-specific results including AF and AN.
}
\description{
Analyze multiple lag structures with AF and AN
}
\keyword{internal}
