% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{attrdl}
\alias{attrdl}
\title{Attrdl
Function for computing attributble measures from dlnm}
\usage{
attrdl(
  x,
  basis,
  cases,
  model = NULL,
  coef = NULL,
  vcov = NULL,
  model.link = NULL,
  type = "af",
  dir = "back",
  tot = TRUE,
  cen,
  range = NULL,
  sim = FALSE,
  nsim = 5000
)
}
\arguments{
\item{x}{An exposure vector or (only for dir="back") a matrix of lagged exposures}

\item{coef, }{vcov: coef and vcov for basis if model is not provided}

\item{basis:}{The cross-basis computed from x}

\item{cases:}{The cases vector or (only for dir="forw") the matrix of future cases}

\item{model:}{The fitted model}

\item{model.link:}{Link function if model is not provided}

\item{type:}{Either "an" or "af" for attributable number or fraction}

\item{dir:}{Either "back" or "forw" for backward or forward perspectives}

\item{tot:}{If true, the total attributable risk is computed}

\item{cen:}{The reference value used as counterfactual scenario}

\item{range:}{The range of exposure. if null, the whole range is used}

\item{sim:}{If simulation samples should be returned. only for tot=true}

\item{nsim:}{Number of simulation samples}
}
\description{
Attrdl
Function for computing attributble measures from dlnm
}
