% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{calculate_air_pollution_af_an}
\alias{calculate_air_pollution_af_an}
\title{Calculate AF and AN for a specific PM2.5 reference value}
\usage{
calculate_air_pollution_af_an(
  data,
  reference = 15,
  var_name = "pm25",
  family = "quasipoisson"
)
}
\arguments{
\item{data}{A region-specific dataset.}

\item{reference}{Integer. PM2.5 reference value. Defaults to 15
(WHO guideline).}

\item{var_name}{Character. PM2.5 variable name. Defaults to "pm25".}

\item{family}{Character. Distribution family for GAM. Defaults to
"quasipoisson".}
}
\value{
List with AF and AN
}
\description{
Calculate AF and AN for a specific PM2.5 reference value
}
\keyword{internal}
