% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{calculate_air_pollution_grid_dims}
\alias{calculate_air_pollution_grid_dims}
\title{Generate a grid size for a certain number of plots.}
\usage{
calculate_air_pollution_grid_dims(n_plots)
}
\arguments{
\item{n_plots}{The number of plots required for the grid.}
}
\value{
A list containing ncol and nrow values for the grid.
}
\description{
Generate a grid size for a certain number of plots.
}
\keyword{internal}
