% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{calculate_min_mortality_temp}
\alias{calculate_min_mortality_temp}
\title{Calculate minimum mortality values}
\usage{
calculate_min_mortality_temp(df_list, regions_df, blup)
}
\arguments{
\item{df_list}{A list of dataframes for each region}

\item{regions_df}{A dataframe with two columns.
Column 1 is abbreviated region names.
Column 2 is full region names.}

\item{blup}{BLUP (best linear unbiased predictions) for an mvmeta model.}
}
\value{
\itemize{
\item argvar: arguments ($fun, $knots, $degree) for cross-basis function
\item bvar:
\item mintempregions: optimum temperature per region
}
}
\description{
???
}
