% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{calculate_qaic}
\alias{calculate_qaic}
\title{QAIC calculation}
\usage{
calculate_qaic(
  data,
  temperature_lag,
  spline_temperature_degrees_freedom,
  calculate_by_region,
  save_csv = FALSE,
  output_folder_path = NULL,
  print_results = FALSE
)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data from which to fit models.}

\item{temperature_lag}{Integer. The number of days for which to calculate
the lags for temperature. Default is 1.}

\item{spline_temperature_degrees_freedom}{Integer. Degrees of freedom for the
spline(s).}

\item{calculate_by_region}{Bool. Whether to calculate Relative Risk
by region. Defaults to FALSE.}

\item{save_csv}{Bool. Whether or not to save the VIF results to a CSV.}

\item{output_folder_path}{String. Where to save the CSV file to (if save_csv == TRUE).}

\item{print_results}{Logical. Whether or not to print model summaries and
pearson dispersion statistics. Defaults to FALSE.}
}
\value{
Dataframe containing QAIC results for each lag.
}
\description{
Computes the Quasi-Akaike Information Criterion (QAIC) for models,
enabling model comparison
}
\keyword{internal}
