% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{casecrossover_dlnm}
\alias{casecrossover_dlnm}
\title{Quasi-Poisson Case-Crossover model with DLNM}
\usage{
casecrossover_dlnm(data, cb_list)
}
\arguments{
\item{data}{Dataframe containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}
}
\value{
List containing models by region
}
\description{
Fits a quasi-Poisson case-crossover with a distributed lag
non-linear model
}
