% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{casecrossover_quasipoisson}
\alias{casecrossover_quasipoisson}
\title{Fit quasipoisson regression models for different lags using a time-stratified
case-crossover approach.}
\usage{
casecrossover_quasipoisson(
  data,
  scale_factor_wildfire_pm = 10,
  temperature_lag,
  spline_temperature_degrees_freedom,
  calculate_by_region,
  save_fig = TRUE,
  output_folder_path = NULL,
  print_model_summaries = TRUE
)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data from which to fit models.}

\item{scale_factor_wildfire_pm}{Numeric. The value to divide the wildfire PM2.5
concentration variables by for alternative interpretation of outputs.
Corresponds to the unit increase in wildfire PM2.5 to give the model
estimates and relative risks (e.g. scale_factor = 10 corresponds to estimates
and relative risks representing impacts of a 10 unit increase in wildfire
PM2.5). Setting this parameter to 0 or 1 leaves the variable unscaled.}

\item{temperature_lag}{Integer. The number of days for which to calculate
the lags for temperature. Default is 1.}

\item{spline_temperature_degrees_freedom}{Integer. Degrees of freedom for the
spline(s).}

\item{calculate_by_region}{Bool. Whether to calculate Relative Risk
by region. Defaults to FALSE.}

\item{save_fig}{Bool. Whether or not to save a figure showing residuals vs
fitted values for each lag. Defaults to FALSE.}

\item{output_folder_path}{String. Where to save the figure. Defaults to NULL.}

\item{print_model_summaries}{Bool. Whether to print the model summaries to
console. Defaults to FALSE.}
}
\value{
Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5
}
\description{
Fits quasipoisson regression models using \link[gnm]{gnm}
}
\keyword{internal}
