% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{check_empty_dataframe}
\alias{check_empty_dataframe}
\title{Check if a dataframe is empty.}
\usage{
check_empty_dataframe(df)
}
\arguments{
\item{df}{Dataframe. The dataframe to check.}
}
\value{
NULL. No return if the dataframe is not empty.
}
\description{
Checks if a dataframe is empty, and raises an error if it is.
}
