% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check that a file exists at a passed path.}
\usage{
check_file_exists(fpath, raise = TRUE)
}
\arguments{
\item{fpath}{The filepath to check exists.}

\item{raise}{Whether or not to raise an error if the file does not exist,
Default: TRUE}
}
\value{
'exists'. Whether or not the file exists on disk.
}
\description{
Checks the files on disk to assert that the passed file is
present.
}
