% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{check_file_extension}
\alias{check_file_extension}
\title{Check that a file extension on a given path matches the expected.}
\usage{
check_file_extension(fpath, expected_ext, param_nm = "fpath", raise = TRUE)
}
\arguments{
\item{fpath}{The filepath.}

\item{expected_ext}{The expected file extension.}

\item{param_nm}{The parameter name that the filepath was passed to
(for error raising), Default: 'fpath'}

\item{raise}{Whether or not to raise an error, Default: TRUE}
}
\value{
Whether or not the passed file has a valid file extension.
}
\description{
This function takes an expected file extension, and validates it
against a user-inputted file path.
}
\keyword{internal}
