% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{check_vif}
\alias{check_vif}
\title{Check variance inflation factors of predictor variables using a linear model}
\usage{
check_vif(data, predictors)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data.}

\item{predictors}{Character vector with each of the predictors to include
in the model. Must contain at least 2 variables.}
}
\value{
Prints variance inflation factors for each predictor variable.
}
\description{
Checks variance inflation factors of predictor variables using a
linear model of the predictor variables on the health outcome.
}
