% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{check_wildfire_vif}
\alias{check_wildfire_vif}
\title{Check variance inflation factors of predictor variables using a linear model}
\usage{
check_wildfire_vif(
  data,
  predictors,
  calculate_by_region,
  join_wildfire_data,
  save_csv = FALSE,
  output_folder_path = NULL,
  print_vif = FALSE
)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data.}

\item{predictors}{Character vector with each of the predictors to include
in the model. Must contain at least 2 variables.}

\item{calculate_by_region}{Bool. Whether to calculate Relative Risk
by region. Defaults to FALSE.}

\item{join_wildfire_data}{Boolean. If TRUE, a daily time series of
wildfire-related PM2.5 concentration is joined to the health data. If FALSE,
the data set is loaded without any additional joins.}

\item{save_csv}{Bool. Whether or not to save the VIF results to a CSV.}

\item{output_folder_path}{String. Where to save the CSV file to (if save_csv == TRUE).}

\item{print_vif}{Bool, whether or not to print VIF for each predictor.
Defaults to FALSE.}
}
\value{
Variance inflation factor statistics for each predictor variable.
}
\description{
Checks variance inflation factors of predictor variables using a
linear model of the predictor variables on the health outcome. Prints stats
if print_vif==TRUE. Raises a warning if VIF for a variables is > 2.
}
\keyword{internal}
