% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{combine_health_climate_data}
\alias{combine_health_climate_data}
\title{Read in and combine climate and health data}
\usage{
combine_health_climate_data(
  health_data_path,
  climate_data_path,
  map_path,
  region_col,
  district_col,
  date_col,
  year_col,
  month_col,
  case_col,
  case_type,
  tot_pop_col,
  tmin_col,
  tmean_col,
  tmax_col,
  rainfall_col,
  r_humidity_col,
  geometry_col,
  runoff_col = NULL,
  ndvi_col = NULL,
  spi_col = NULL,
  max_lag,
  output_dir = NULL
)
}
\arguments{
\item{health_data_path}{The path to the health data.}

\item{climate_data_path}{The path to the climate data.}

\item{map_path}{The path to the relevant map data.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names.}

\item{district_col}{Character. Name of the column in the dataframe that
contains the region names.}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date. Defaults to NULL.}

\item{year_col}{Character. Name of the column in the dataframe that contains
the Year.}

\item{month_col}{Character. Name of the column in the dataframe that contains
the Month.}

\item{case_col}{Character. Name of the column in the dataframe
that contains the disease cases to be considered.}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of 'diarrhea' or 'malaria'.}

\item{tot_pop_col}{Character. Name of the column in the dataframe that
contains the total population.}

\item{tmin_col}{Character. Name of the column in the dataframe that
contains the minimum temperature data.}

\item{tmean_col}{Character. Name of the column in the dataframe that
contains the average temperature.}

\item{tmax_col}{Character. Name of the column in the dataframe that
contains the maximum temperature.}

\item{rainfall_col}{Character. Name of the column in the dataframe that
contains the cumulative monthly rainfall.}

\item{r_humidity_col}{Character. Name of the column in the dataframe that
contains the relative humidity.}

\item{geometry_col}{is the Name of the geometry column in the shapefile
(usually "geometry").}

\item{runoff_col}{Character. Name of the column in the dataframe that
contains the monthly runoff water data. Defaults to NULL.}

\item{ndvi_col}{Character. Name of column containing the Normalized Difference
Vegetation Index (ndvi) data. Defaults to NULL.}

\item{spi_col}{Character. Name of the column in the dataframe that
contains the standardized precipitation index. Defaults to NULL.}

\item{max_lag}{Character. Number corresponding to the maximum lag to be
considered for the delay effect. It should be between 2 an 4.}

\item{output_dir}{Path to folder where the processed map data should be
saved. Defaults to NULL.}
}
\value{
A list of dataframes containing the map, nb.map, data, grid_data, summary
}
\description{
Read and combine climate and health data prepared for the
spatiotemporal and DLNM analysis.
}
\keyword{internal}
