% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{common_descriptive_stats}
\alias{common_descriptive_stats}
\title{The wrapper function to compute descriptive stats for the heat and cold indicator.}
\usage{
common_descriptive_stats(
  dataset_title,
  df_list,
  output_path,
  plot_corr_matrix = F,
  correlation_method = "pearson",
  plot_dist = F,
  dist_columns = c(),
  plot_ma = F,
  ma_days = 100,
  ma_sides = 2,
  ma_columns = c(),
  timeseries_col = NULL,
  dependent_col,
  independent_cols,
  plot_na_counts = F,
  plot_scatter = F
)
}
\arguments{
\item{dataset_title}{Character. The title for the dataset being used.}

\item{df_list}{Vector. A list of input dataframes.}

\item{output_path}{Character. The path to write outputs to.}

\item{plot_corr_matrix}{Bool. Whether or not to plot a correlation matrix. Defaults to F.}

\item{correlation_method}{Character. The correlation method used in ds.}

\item{plot_dist}{Bool. Whether or not to plot distribution histograms. Defaults to F.}

\item{dist_columns}{Vector. The columns to plot distributions for. Defaults to c().}

\item{plot_ma}{Bool. Whether or not to plot moving average for specified columns. Defaults to F.}

\item{ma_days}{Integer. The number of days to use for a moving average. Defaults to 100.}

\item{ma_sides}{Integer. The number of sides to use for a moving average (1 or 2). Defaults to 2.}

\item{ma_columns}{Vector. Additional columns to plot moving average for. Defaults to c().}

\item{timeseries_col}{Character. The column containing the timeseries (e.g., Date). Defaults to NULL.}

\item{dependent_col}{Character. The column in the data containing the dependent var.}

\item{independent_cols}{Character. The column in the data containing the independent var.}

\item{plot_na_counts}{Bool. Whether to plot NA counts. Defaults to F.}

\item{plot_scatter}{Bool. Whether to plot a scatter plot of independent vs dependent. Defaults to F.}
}
\description{
The wrapper function to compute descriptive stats for the heat and cold indicator.
}
