% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{common_descriptive_stats_core}
\alias{common_descriptive_stats_core}
\title{The core functionality for producing descriptive stats}
\usage{
common_descriptive_stats_core(
  dataset_title,
  df,
  output_path,
  title,
  plot_corr_matrix = F,
  correlation_method = "pearson",
  plot_dist = F,
  dist_columns = c("temp"),
  dependent_col = "dependent",
  independent_cols = c("temp"),
  plot_na_counts = F,
  plot_scatter = F
)
}
\arguments{
\item{dataset_title}{Character. The title for the dataset being used.}

\item{df}{Dataframe. The input DataFrame.}

\item{output_path}{Character. The path to write outputs to.}

\item{title}{Character. The specific title for the subset of data being used.}

\item{plot_corr_matrix}{Bool. Whether or not to plot a correlation matrix. Defaults to F.}

\item{correlation_method}{Character. The correlation method used in ds.}

\item{plot_dist}{Bool. Whether or not to plot distribution histograms. Defaults to F.}

\item{dist_columns}{Vector. The columns to plot distributions for. Defaults to c().}

\item{dependent_col}{Character. The column in the data containing the dependent var.}

\item{independent_cols}{Character. The column in the data containing the independent var.}

\item{plot_na_counts}{Bool. Whether to plot NA counts. Defaults to F.}

\item{plot_scatter}{Bool. Whether to plot a scatter plot of independent vs dependent. Defaults to F.}
}
\description{
The core functionality for producing descriptive stats
}
