% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{compute_attributable_deaths}
\alias{compute_attributable_deaths}
\title{Compute attributable deaths}
\usage{
compute_attributable_deaths(
  df_list,
  output_year,
  blup = NULL,
  mintempregions,
  an_thresholds,
  independent_cols = NULL,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas,
  nsim_ = 1000
)
}
\arguments{
\item{df_list}{An alphabetically-ordered list
of dataframes for each region.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{mintempregions}{A named numeric vector.
Minimum (optimum) mortality temperature per region.}

\item{an_thresholds}{A dataframe with the optimal temperature range and
temperature thresholds for calculation of attributable deaths.}

\item{independent_cols}{column name (or list of names) of extra independent
variable to include in regression (excluding temperature). Defaults to NULL.}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}

\item{nsim_}{The number of simulation runs used for computing emprical CI}
}
\value{
A list of variables
\itemize{
\item \code{totdeath} A named vector of integers.
otal observed mortality per region.
\item \code{arraysim} An array (numeric). Total (glob),
cold and heat-attributable deaths per region for 1000 simulations.
Used to derive confidence intervals.
\item \code{matsim} A matrix (numeric). Total (glob),
cold and heat-attributable deaths per region from reduced coefficients.
\item \code{attrdl_yr_all} a dataframe containing attributable deaths by year
for each region.
}
}
\description{
Compute the attributable deaths for each regions,
with empirical CI estimated using the re-centered bases.
}
