% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{compute_attributable_rates}
\alias{compute_attributable_rates}
\title{Compute attributable rates}
\usage{
compute_attributable_rates(df_list, output_year, matsim, arraysim)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each
region comprising dates, deaths, and temperatures.}

\item{output_year}{Year(s) to calculate output for.}

\item{matsim}{A matrix (numeric). Total (glob),
cold and heat-attributable deaths per region from reduced coefficients.}

\item{arraysim}{An array (numeric). Total (glob),
cold and heat-attributable deaths per region for 1000 simulations.
Used to derive confidence intervals.}
}
\value{
\itemize{
\item \code{anregions_bind}
\item \code{antot_bind}
\item \code{arregions_bind}
\item \code{artot_bind}
}
}
\description{
Compute attributable rates
}
