% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{create_air_pollution_lags}
\alias{create_air_pollution_lags}
\title{Create lagged values for PM2.5 variable and average lag column.}
\usage{
create_air_pollution_lags(data, max_lag = 14L)
}
\arguments{
\item{data}{Dataframe from  load_air_pollution_data() containing a daily
time series of health and environmental data.}

\item{max_lag}{Integer. The maximum lag days for outdoor PM2.5. Defaults to 14.}
}
\value{
Dataframe with added columns for lagged PM2.5 concentration.
}
\description{
Creates new variables in a dataframe for lags and means over lag
periods.
}
\keyword{internal}
