% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{create_column_summaries}
\alias{create_column_summaries}
\title{Create statistical summaries of columns in a dataframe.}
\usage{
create_column_summaries(df, independent_cols = NULL)
}
\arguments{
\item{df}{Datarame. Input data.}

\item{independent_cols}{Character vector. The columns in the data containing the independent variables.}
}
\value{
Dataframe. Column summaries
}
\description{
Create statistical summaries of columns in a dataframe.
}
\keyword{internal}
