% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{create_correlation_matrix}
\alias{create_correlation_matrix}
\title{Create a correlation matrix for columns in a dataframe.}
\usage{
create_correlation_matrix(df, columns = NULL, correlation_method = "pearson")
}
\arguments{
\item{df}{Dataframe. The dataframe to use to create a correlation matrix.}

\item{columns}{vector. The columns to calculate correlation between.}

\item{correlation_method}{string. The method to use for correlation calculations.}
}
\value{
Matrix. Correlation matrix for selected columns in the input dataset.
}
\description{
Create a correlation matrix for columns in a dataframe.
}
\keyword{internal}
