% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{create_crossbasis}
\alias{create_crossbasis}
\title{Create cross-basis matrix}
\usage{
create_crossbasis(data, var_fun, var_dof, lag_fun, lag_dof, lag_days)
}
\arguments{
\item{data}{Dataframe containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_fun}{Character. Exposure function
(see dlnm::crossbasis).}

\item{var_dof}{Integer. Degrees of freedom in exposure function
(see dlnm:crossbasis).}

\item{lag_fun}{Character. Exposure function
(see dlnm::crossbasis).}

\item{lag_dof}{Integer. Degrees of freedom in exposure function
(see dlnm:crossbasis).}

\item{lag_days}{Integer. Maximum lag
(see dlnm:crossbasis).}
}
\value{
Large list of cross-basis matrices by region
}
\description{
Creates a cross-basis matrix for each region
}
