% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarrhea.R
\name{create_inla_indices}
\alias{create_inla_indices}
\title{Create indices for INLA models}
\usage{
create_inla_indices(data)
}
\arguments{
\item{data}{is the dataset containing district_code, region_code, and year
columns from the combine_health_climate_data() function.}
}
\value{
The modified data with the created indices.
}
\description{
: For the INLA model, there is a need to set-up regions index,
district index, and year index. This function create these indices using the
dataset, ndistrict and nregion defined above.
}
