% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{create_lagged_variables}
\alias{create_lagged_variables}
\title{Generate lagged values for predictor variables}
\usage{
create_lagged_variables(data, wildfire_lag = 3, temperature_lag = 1)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data}

\item{wildfire_lag}{Integer. The number of days for which to calculate the
lags for wildfire PM2.5. Default is 3.}

\item{temperature_lag}{Integer. The number of days for which to calculate
the lags for temperature. Default is 1.}
}
\value{
Dataframe with added columns for lagged temperature and
wildfire-related PM2.5 concentration
}
\description{
Generates new variables in a dataframe for lags and means over
lag periods.
}
