% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{create_na_summary}
\alias{create_na_summary}
\title{Create a summary of all NA values in a dataset.}
\usage{
create_na_summary(df, independent_cols = NULL)
}
\arguments{
\item{df}{Dataframe. The input dataset.}

\item{independent_cols}{Character vector. The columns in the data containing the independent variables.}
}
\value{
Dataframe. A summary of NA values in the dataset.
}
\description{
Create a summary of all NA values in a dataset.
}
\keyword{internal}
