% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{define_and_validate_optimal_temps}
\alias{define_and_validate_optimal_temps}
\title{Define and validate the optimal temperature range from the model predictions.}
\usage{
define_and_validate_optimal_temps(
  optimal_temp_range,
  prediction,
  RR_fit_col = "allRRfit",
  index
)
}
\arguments{
\item{optimal_temp_range}{Matrix. A matrix used to store the optimal temperature
ranges.}

\item{prediction}{Data. The models prediction.}

\item{RR_fit_col}{Character. The column containing the relative risk values.}

\item{index}{Integer. The index to use to obtain the RR values.}
}
\value{
The optimal temperature range.
}
\description{
Define and validate the optimal temperature range from the model predictions.
}
