% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{define_model}
\alias{define_model}
\title{Define regression model}
\usage{
define_model(dataset)
}
\arguments{
\item{data}{dataframe with tmean column to be modelled}
}
\value{
\itemize{
\item \code{model} A quasi-poission generalised linear model object.
See: https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/glm
\item \code{cb} Basis matrices for the two dimensions of predictor and lags.
}
}
\description{
Define regression model
}
