% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Show descriptive statistics}
\usage{
descriptive_stats(data, variables, bin_width = 5)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data}

\item{variables}{Character or character vector with variable to produce
summary statistics for. Must include at least 1 variable.}

\item{bin_width}{Integer. Width of each bin in a histogram of the outcome
variable. Default is 5.}
}
\value{
Prints summary statistics and a histogram of the the outcome
variable
}
\description{
Generates summary statistics for climate and health data.
}
\keyword{internal}
