% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect Outliers Using the IQR Method}
\usage{
detect_outliers(df, independent_cols = NULL)
}
\arguments{
\item{df}{A data frame containing the data to check for outliers.}

\item{independent_cols}{Character vector. The columns in the data containing the independent variables.}
}
\value{
Dataframe. Column summaries
}
\description{
Detect Outliers Using the IQR Method
}
\keyword{internal}
