% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnm_shared.R
\name{dlnm_meta_analysis}
\alias{dlnm_meta_analysis}
\title{Meta-analysis and BLUPs}
\usage{
dlnm_meta_analysis(
  df_list,
  coef_,
  vcov_,
  save_csv = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{coef_}{A matrix of coefficients for the reduced model.}

\item{vcov_}{A list. Covariance matrices for each region for the reduced model.}

\item{save_csv}{Boolean. Whether to save the results as a CSV. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where results should be saved.
Defaults to NULL.}
}
\value{
\itemize{
\item \code{mm} A model object. A multivariate meta-analysis model.
\item \code{blup} A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each region.
\item \code{meta_test_res} A dataframe of results from statistical tests on the meta model.
}
}
\description{
Run meta-analysis using temperature average and range as meta
predictors. Then create the best linear unbiased predictions (BLUPs).
}
\keyword{internal}
