% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnm_shared.R
\name{dlnm_min_mortality_temp}
\alias{dlnm_min_mortality_temp}
\title{Define minimum mortality percentiles and temperatures}
\usage{
dlnm_min_mortality_temp(
  df_list,
  var_fun = "bs",
  var_per = c(10, 75, 90),
  var_degree = 2,
  blup = NULL,
  coef_,
  meta_analysis = FALSE,
  outcome_type = c("temperature", "suicide")
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(10, 75, 90).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{blup}{A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each region. Defaults to NULL.}

\item{coef_}{A matrix of coefficients for the reduced model.}

\item{meta_analysis}{Boolean. Whether to perform a meta-analysis. Defaults
to FALSE.}

\item{outcome_type}{Character. The indicator that the function is being used
for. One of 'suicide' or 'temperature'. Defaults to c("temperature", "suicide")}
}
\value{
Percentiles and corresponding temperatures for each geography.
}
\description{
Calculate the temperature at which there is minimum mortality risk
using the product of the basis matrix and BLUPs.
}
\keyword{internal}
