% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnm_shared.R
\name{dlnm_power_list}
\alias{dlnm_power_list}
\title{Power calculation}
\usage{
dlnm_power_list(
  df_list,
  pred_list,
  minperc,
  attr_thr_high = 97.5,
  attr_thr_low = 2.5,
  compute_low = TRUE
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{attr_thr_high}{Integer. Percentile at which to define the upper temperature threshold for
calculating attributable risk. Defaults to 97.5.}

\item{minpercreg}{Vector. Percentile of maximum outcome temperature for each region.}

\item{attr_thr_lower}{Integer. Percentile at which to define the lower temperature threshold for
calculating attributable risk. Defaults to 2.5.}

\item{compute_lowe}{Bool. Whether to computer power for the lower threshold.
Defaults to FALSE}
}
\value{
A list containing power information by area.
}
\description{
Produce a power statistic by area for the attributable threshold
and above as a reference.
}
\keyword{internal}
