% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnm_shared.R
\name{dlnm_reduce_cumulative}
\alias{dlnm_reduce_cumulative}
\title{Reduce to overall cumulative}
\usage{
dlnm_reduce_cumulative(
  df_list,
  var_per = c(25, 50, 75),
  var_degree = 2,
  cenper = NULL,
  cb_list,
  model_list
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(25, 50, 75).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{cenper}{Integer. Value for the percentile in calculating the centering
value 0-100. Defaults to 50.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}

\item{model_list}{List of models produced from case-crossover and DLNM
analysis.}
}
\value{
\itemize{
\item \code{coef_} A matrix of coefficients for the reduced model.
\item \code{vcov_} A list. Covariance matrices for each region for the reduced model.
}
}
\description{
Reduce model to the overall cumulative association
}
\keyword{internal}
