% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnm_shared.R
\name{dlnm_vif}
\alias{dlnm_vif}
\title{Produce variance inflation factor}
\usage{
dlnm_vif(df_list, independent_cols = NULL)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{independent_cols}{Additional independent variables to test in model validation.
Defaults to NULL.}
}
\value{
A list. Variance inflation factors for each independent variables by region.
}
\description{
Produces variance inflation factor for the independent variables.
}
\keyword{internal}
