% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{enforce_file_extension}
\alias{enforce_file_extension}
\title{Enforce a file extension on a given path}
\usage{
enforce_file_extension(path, file_extension)
}
\arguments{
\item{path}{Character. A file path.}

\item{file_extension}{Character. The file extension to enforce on 'path'.}
}
\value{
Character. The path with the expected file extension.
}
\description{
Ensures that the provided file path ends with the desired file extension.
}
\examples{
enforce_file_extension("output", "pdf")
enforce_file_extension("report.txt", ".pdf")

}
\keyword{internal}
