% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{extract_air_pollution_coef}
\alias{extract_air_pollution_coef}
\title{Extract coefficient and standard error from GAM model}
\usage{
extract_air_pollution_coef(model, var_name = "pm25")
}
\arguments{
\item{model}{GAM model object from single region.}

\item{var_name}{Character. Column name to extract. Defaults to "pm25"}
}
\value{
List with coefficient and standard error.
}
\description{
Extracts the coefficient and standard error of the GAM model.
}
\keyword{internal}
