% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{filter_on_rr_distribution}
\alias{filter_on_rr_distribution}
\title{Filter the dataframe based on the relative risk distribution.}
\usage{
filter_on_rr_distribution(
  df,
  RR_distribution_length = 0,
  lower_range = 5,
  upper_range = 200,
  output_year = 0
)
}
\arguments{
\item{df}{The dataframe to filter.}

\item{RR_distribution_length}{The RR distribution length. Defaults to 0.}

\item{lower_range}{The lower range. Defaults to 5.}

\item{upper_range}{The upper range. Defaults to 15.}

\item{output_year}{The output year. Defaults to 0.}
}
\value{
The filtered dataframe.
}
\description{
Filter the dataframe based on the relative risk distribution.
}
