% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{generate_descriptive_stats_run_id}
\alias{generate_descriptive_stats_run_id}
\title{Generate a run id for descriptive statistics output folders.}
\usage{
generate_descriptive_stats_run_id()
}
\value{
Character. Run id in the format YYYYmmdd_HHMMSS_NNNN.
}
\description{
Generate a run id for descriptive statistics output folders.
}
\keyword{internal}
