% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accessibility_helpers.R
\name{get_accessible_ggplot_grid}
\alias{get_accessible_ggplot_grid}
\title{Get accessible ggplot grid dimensions}
\usage{
get_accessible_ggplot_grid(n_plots)
}
\arguments{
\item{n_plots}{Integer. Number of panels or plots.}
}
\value{
A list with \code{n_col} and \code{n_row}.
}
\description{
Calculates grid dimensions for ggplot facets or patchwork plots. Defaults to
a maximum of two columns to avoid squished panels.
}
\keyword{internal}
