% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{get_alpha_colour}
\alias{get_alpha_colour}
\title{Generate and RGB colour value with alpha from a hex value.}
\usage{
get_alpha_colour(hex, alpha)
}
\arguments{
\item{hex}{The hex code of the colour to convert.}

\item{alpha}{The alpha of the converted colour (ranging from 0-1).}
}
\value{
The converted RGB colour.
}
\description{
Generate and RGB colour value with alpha from a hex value.
}
