% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{get_lags_and_means}
\alias{get_lags_and_means}
\title{Create lagged columns and provide the mean value.}
\usage{
get_lags_and_means(data, lagcol, nlags)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data}

\item{lagcol}{Character. The column to lag.}

\item{nlags}{Character. How many rows to obtain a lag from.}
}
\value{
Dataframe with added columns for lagged values and mean(s) of those
lags.
}
\description{
Creates new columns containing lagged values over n rows and
determine the mean of the lagged column.
}
\keyword{internal}
