% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accessibility_helpers.R
\name{get_layout_matrix}
\alias{get_layout_matrix}
\title{Get layout matrix for multi-panel plot}
\usage{
get_layout_matrix(n_plots, n_col, n_row)
}
\arguments{
\item{n_plots}{Integer. Number of plots.}

\item{n_col}{Integer. Number of columns.}

\item{n_row}{Integer. Number of rows.}
}
\value{
A matrix for \code{layout()}.
}
\description{
Creates a layout matrix for use with \code{layout()}.
}
\keyword{internal}
