% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_create_crossbasis}
\alias{hc_create_crossbasis}
\title{Create cross-basis matrix}
\usage{
hc_create_crossbasis(
  df_list,
  var_fun = "bs",
  var_degree = 2,
  var_per = c(10, 75, 90),
  lagn = 21,
  lagnk = 3
)
}
\arguments{
\item{df_list}{A list of data frames containing daily time series data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(10,75,90).}

\item{lagn}{Integer. Number of days in the lag period. Defaults to 21.
(see dlnm::crossbasis).}

\item{lagnk}{Integer. Number of knots in lag function. Defaults to 3.
(see dlnm::logknots).}
}
\value{
'cb_list'. A list of cross-basis matrices by geography.
}
\description{
Creates a cross-basis matrix of the lag-response and
exposure-response functions, for each geography.
}
\keyword{internal}
