% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_model_validation}
\alias{hc_model_validation}
\title{Model Validation Assessment}
\usage{
hc_model_validation(
  df_list,
  cb_list,
  independent_cols = NULL,
  dfseas = 8,
  save_fig = FALSE,
  save_csv = FALSE,
  output_folder_path = NULL,
  seed = NULL
)
}
\arguments{
\item{df_list}{A list of data frames containing daily time series data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{cb_list}{List of cross-basis matrices from hc_create_crossbasis
function.}

\item{independent_cols}{Character/list. Additional independent variables to
test in model validation as confounders. Defaults to NULL.}

\item{dfseas}{Integer. Degrees of freedom for seasonality. Defaults to 8.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{save_csv}{Boolean. Whether to save the results as a CSV. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}
}
\value{
\itemize{
\item \code{qaic_results} A data frame of QAIC and dispersion metrics for each
model combination and geography.
\item \code{qaic_summary} A data frame with the mean QAIC and dispersion metrics
for each model combination.
\item \code{vif_results} A data frame of variance inflation factors for each
independent variables by geography.
\item \code{vif_summary} A data frame with the mean variance inflation factors
for each independent variable.
\item \code{adf_results} A data frame of ADF test results for each geography.
}
}
\description{
Produces results on QAIC for each model combination, variance
inflation factor for each independent variable, ADF test for stationarity,
and plots for residuals to assess the models.
}
\keyword{internal}
