% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_plot_af_heat_monthly}
\alias{hc_plot_af_heat_monthly}
\title{Plot attributable fractions by calendar month - high temperatures}
\usage{
hc_plot_af_heat_monthly(
  attr_mth_list,
  df_list,
  country = "National",
  attr_thr_high = 97.5,
  save_fig = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{attr_mth_list}{A list of data frames containing total attributable
fractions, numbers and rates by calendar month and geography.}

\item{df_list}{A list of data frames containing daily time series data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{country}{Character. Name of country for national level estimates.
Defaults to 'National'.}

\item{attr_thr_high}{Integer. Percentile at which to define the upper
temperature threshold for calculating attributable risk. Defaults to 97.5.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}
}
\value{
Plots of attributable fractions by calendar month per geography.
}
\description{
Plot attributable fractions grouped over the whole time series
by calendar month to explore seasonality.
}
\keyword{internal}
