% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_plot_ar_heat_yearly}
\alias{hc_plot_ar_heat_yearly}
\title{Plot attributable rates by year - high temps}
\usage{
hc_plot_ar_heat_yearly(
  attr_yr_list,
  save_fig = FALSE,
  output_folder_path = NULL,
  country = "National"
)
}
\arguments{
\item{attr_yr_list}{A list of matrices containing yearly estimates of attributable
fractions, numbers and rates by area}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}

\item{country}{Character. Name of country for national level estimates.}
}
\value{
Plots of yearly attributable rates per area
}
\description{
Plot attributable rates by year and area with confidence intervals
}
\keyword{internal}
