% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_plot_attr_cold_totals}
\alias{hc_plot_attr_cold_totals}
\title{Plot total attributable fractions and rates - low temperatures}
\usage{
hc_plot_attr_cold_totals(
  df_list,
  res_attr_tot,
  save_fig = FALSE,
  output_folder_path = NULL,
  country = "National"
)
}
\arguments{
\item{df_list}{A list of data frames containing daily time series data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{res_attr_tot}{Matrix containing total attributable fractions, numbers
and rates for each geography over the whole time series.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}

\item{country}{Character. Name of country for national level estimates.
Defaults to 'National'.}
}
\value{
Plots of total attributable fractions and rates by geography
}
\description{
Plot total attributable fractions and rates over the whole
time series by geography.
}
\keyword{internal}
