% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_plot_power}
\alias{hc_plot_power}
\title{Plot statistical power for temperature mortality analysis}
\usage{
hc_plot_power(
  power_list_high,
  power_list_low,
  save_fig = FALSE,
  output_folder_path = NULL,
  country = "National"
)
}
\arguments{
\item{power_list_high}{List. A list containing power information for high
temperatures by geography.}

\item{power_list_low}{List. A list containing power information for low
temperatures by geography.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults
to FALSE.}

\item{output_folder_path}{Character. Path to folder where plots should be
saved. Defaults to NULL.}

\item{country}{Character. The name of the country for national level
estimates. Defaults to 'National'.}
}
\value{
Plots of power by temperature for the attributable thresholds
and above for each geography.
}
\description{
Plots the power statistic for each reference temperature at
and above the attributable risk threshold for each geography.
}
\keyword{internal}
