% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_predict_subnat}
\alias{hc_predict_subnat}
\title{Run predictions from model}
\usage{
hc_predict_subnat(
  df_list,
  var_fun = "bs",
  var_per = c(10, 75, 90),
  var_degree = 2,
  mintempgeog_,
  blup,
  coef_,
  vcov_,
  meta_analysis = FALSE
)
}
\arguments{
\item{df_list}{A list of data frames containing daily time series data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(10, 75, 90).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{mintempgeog_}{Vector. Percentile of minimum mortality temperature for
each geography.}

\item{blup}{A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each geography.}

\item{coef_}{A matrix of coefficients for the reduced model.}

\item{vcov_}{A list. Covariance matrices for each geography for the reduced
model.}

\item{meta_analysis}{Boolean. Whether to perform a meta-analysis. Defaults
to FALSE.}
}
\value{
'pred_list'. A list containing predictions by geography.
}
\description{
Use model to run predictions. Predictions can be produced for a
single input geography, or multiple disaggregated geographies.
}
\keyword{internal}
