% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_read_data}
\alias{hc_read_data}
\title{Read temperature-related mortality indicator data}
\usage{
hc_read_data(
  input_csv_path,
  dependent_col,
  date_col,
  region_col,
  temperature_col,
  population_col
)
}
\arguments{
\item{input_csv_path}{Path to a CSV containing a daily time series of health
outcome and climate data per geography.}

\item{dependent_col}{Character. Name of the column in the data frame
containing the dependent health outcome variable e.g. deaths.}

\item{date_col}{Character. Name of the column in the data frame containing
the date.}

\item{region_col}{Character. Name of the column in the data frame that
contains the geography name(s).}

\item{temperature_col}{Character. Name of the column in the data frame that
contains the temperature column.}

\item{population_col}{Character. Name of the column in the data frame that
contains the population estimate per geography.}
}
\value{
'df_list'. A list of data frames for each geography with formatted
and renamed columns.
}
\description{
Reads in data and geography names for analysis from a CSV file.
}
\keyword{internal}
