% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_rr_results}
\alias{hc_rr_results}
\title{Produce cumulative relative risk results of analysis}
\usage{
hc_rr_results(
  pred_list,
  df_list,
  minpercgeog_,
  attr_thr_high = 97.5,
  attr_thr_low = 2.5
)
}
\arguments{
\item{pred_list}{A list containing predictions from the model by geography.}

\item{df_list}{A list of data frames containing daily time series data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{minpercgeog_}{Vector. Percentile of minimum mortality temperature for
each geography.}

\item{attr_thr_high}{Integer. Percentile at which to define the upper
temperature threshold for calculating attributable risk. Defaults to 97.5.}

\item{attr_thr_low}{Integer. Percentile at which to define the lower
temperature threshold for calculating attributable risk. Defaults to 2.5.}
}
\value{
'rr_results'. Data frame containing cumulative relative risk and
confidence intervals from analysis.
}
\description{
Produces cumulative relative risk and confidence intervals
from analysis.
}
\keyword{internal}
