% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{heat_and_cold_descriptive_stats}
\alias{heat_and_cold_descriptive_stats}
\title{The wrapper function to compute descriptive stats for the heat and cold indicator.}
\usage{
heat_and_cold_descriptive_stats(
  df_list,
  use_individual_dfs,
  output_path,
  correlation_method = "pearson",
  dist_columns,
  ma_days = 100,
  ma_sides = 2,
  dependent_col,
  independent_cols
)
}
\arguments{
\item{df_list}{list(dataframe) A list of input dataframes.}

\item{use_individual_dfs}{bool. Whether or not to use all dfs or concat them into one.}

\item{output_path}{string. The path to write outputs to.}

\item{correlation_method}{string. The correlation method used in ds.}

\item{dist_columns}{vector. The columns to plot distributions for.}

\item{ma_days}{int. The number of days to use for a moving average.}

\item{ma_sides}{int. The number of sides to use for a moving average (1 or 2).}

\item{dependent_col}{str. The column in the data containing the dependent var.}

\item{independent_cols}{str. The column in the data containing the independent var.}

\item{title}{string. title used for the plots in the ds.}
}
\description{
The wrapper function to compute descriptive stats for the heat and cold indicator.
}
