% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defence_utils.R
\name{is_list_of_dfs}
\alias{is_list_of_dfs}
\title{Check that a passed argument is a list of dataframes.}
\usage{
is_list_of_dfs(list_, raise = TRUE, param_nm = "df_list")
}
\arguments{
\item{list_}{A list of items.}

\item{raise}{Whether or not to raise an error.}

\item{param_nm}{The name of the parameter that the list was originally passed
to.}
}
\value{
Whether or not the item passed (list_) is a list of dfs.
}
\description{
Check that a passed argument is a list of dataframes.
}
\keyword{internal}
