% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{join_health_and_climate_data}
\alias{join_health_and_climate_data}
\title{Join health and climate data}
\usage{
join_health_and_climate_data(
  climate_data,
  health_data,
  region_col = "region",
  date_col = "date",
  exposure_col = "mean_PM"
)
}
\arguments{
\item{climate_data}{Character. Dataframe containing a daily time series of
climate data, which may be disaggregated by region.}

\item{health_data}{Character. Path to a CSV file containing a daily time
series of data for a particular health outcome, which may be disaggregated by
region.}

\item{region_col}{Character. Name of the region column in both datasets.
Defaults to 'region'}

\item{date_col}{Character. Name of the date column in both datasets.
Defaults to 'date'}

\item{exposure_col}{Character. Name of the column in the climate data
containing PM2.5 values in micrograms. Defaults to 'mean_PM'.}
}
\value{
Dataframe containing a daily time series of the joined
climate and health data.
}
\description{
Joins a daily time series of wildfire PM2.5 data with a daily time
series of health data.
}
\keyword{internal}
