% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{label_with_unit}
\alias{label_with_unit}
\title{Append the units to a column label.}
\usage{
label_with_unit(col, units)
}
\arguments{
\item{col}{Character. The column name.}

\item{units}{Named Character vector. A vector of units (str) that map to
columns.}
}
\value{
The new column label containing units (if col in units).
}
\description{
Append the units to a column label.
}
