% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{load_and_process_climatedata}
\alias{load_and_process_climatedata}
\title{Read in and format climate data}
\usage{
load_and_process_climatedata(
  climate_data_path,
  district_col,
  year_col,
  month_col,
  tmin_col,
  tmean_col,
  tmax_col,
  rainfall_col,
  r_humidity_col,
  runoff_col = NULL,
  ndvi = NULL,
  spi_col = NULL,
  max_lag = 4
)
}
\arguments{
\item{climate_data_path}{Path to a csv file containing a monthly time series of data
for climate variables, which may be disaggregated by district.}

\item{district_col}{Character. Name of the column in the dataframe that
contains the region names.}

\item{year_col}{Character. Name of the column in the dataframe that
contains the Year.}

\item{month_col}{Character. Name of the column in the dataframe that
contains the month.}

\item{tmin_col}{Character. Name of the column in the dataframe that
contains the minimum temperature data.}

\item{tmean_col}{Character. Name of the column in the dataframe that
contains the average temperature.}

\item{tmax_col}{Character. Name of the column in the dataframe that
contains the maximum temperature.}

\item{rainfall_col}{Character. Name of the column in the dataframe that
contains the cumulative monthly rainfall.}

\item{r_humidity_col}{Character. Name of the column in the dataframe that
contains the relative humidity.}

\item{runoff_col}{Character. Name of the column in the dataframe that
contains the monthly runoff water data. Defaults to NULL.}

\item{spi_col}{Character. Name of the column in the dataframe that
contains the standardized precipitation index. Defaults to NULL.}

\item{max_lag}{Character. Number corresponding to the maximum lag to be
considered for the delay effect. It should be between 2 an 4. Defaults to 4.}

\item{ndvi_col}{Character. Name of column containing the Normalized Difference
Vegetation Index (ndvi) data. Defaults to NULL.}
}
\value{
climate dataframe with formatted and renamed columns, and the lag
variables
}
\description{
Read in a monthly time series of climate data, rename
columns and create lag variable for spatiotemporal and DLNM analysis. The
climate data should start a year before a start year in the health data to
allow the lag variables calculation.
}
\keyword{internal}
