% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarrhea.R
\name{load_and_process_data}
\alias{load_and_process_data}
\title{Read in and format health data - Diarrhea diseases case}
\usage{
load_and_process_data(
  health_data_path,
  region_col,
  district_col,
  date_col = NULL,
  year_col = NULL,
  month_col = NULL,
  diarrhea_case_col,
  tot_pop_col
)
}
\arguments{
\item{health_data_path}{Path to a csv file containing a monthly time series of data
for Diarrhea outcome, which may be disaggregated by sex (under five case or
above five case), and by Region and District.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names.}

\item{district_col}{Character. Name of the column in the dataframe that
contains the district names.}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date. Defaults to NULL.}

\item{year_col}{Character. Name of the column in the dataframe that contains
the year.}

\item{month_col}{Character. Name of the column in the dataframe that contains
the month.}

\item{diarrhea_case_col}{Character. Name of the column in the dataframe
that contains the Diarrhea cases to be considered.}

\item{tot_pop_col}{Character. Name of the column in the dataframe that contains
the total population.}
}
\value{
A dataframe with formatted and renamed columns.
}
\description{
Reads in a csv file containing a monthly time series of health
outcomes and population data. Renames columns and creates time variables for
spatiotemporal analysis.
}
