% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{load_and_process_map}
\alias{load_and_process_map}
\title{Read in and format country map data}
\usage{
load_and_process_map(
  map_path,
  region_col,
  district_col,
  geometry_col,
  output_dir = NULL
)
}
\arguments{
\item{map_path}{The path to the country's shape file "sf" data.}

\item{region_col}{Character. The region column in the dataset.}

\item{district_col}{Character. The district column in the dataset.}

\item{geometry_col}{Character. The geometry column in the dataset.}

\item{output_dir}{Character. The path to output the processed map data to.}
}
\value{
\itemize{
\item 'map' The processed map
\item 'nb.map'
\item 'graph_file'
}
}
\description{
: Read in a shape file, rename columns and create the
adjacency matrix for spatiotemporal analysis.
}
\keyword{internal}
