% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{load_data}
\alias{load_data}
\title{Load data for analysis}
\usage{
load_data(input_path)
}
\arguments{
\item{input_path}{Path to a CSV contain daily time series of death and temperature per region.}
}
\value{
\itemize{
\item df_list_unordered A list of dataframes for each region comprising dates, deaths, and temperatures.
\item regions A list of strings of the names of each region.
}
\description{
Load data for analysis
}
