% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{load_wildfire_data}
\alias{load_wildfire_data}
\title{Load wildfire and health data}
\usage{
load_wildfire_data(
  health_path,
  join_wildfire_data,
  ncdf_path,
  shp_path,
  date_col,
  region_col = NULL,
  mean_temperature_col,
  health_outcome_col,
  pm_2_5_col = NULL
)
}
\arguments{
\item{health_path}{Path to a CSV file containing a
daily time series of data for a particular health outcome, which may be
disaggregated by region. If this does not include a column with
wildfire-related PM2.5, use join_wildfire_data = TRUE to join these data.}

\item{join_wildfire_data}{Boolean. If TRUE, a daily time series of
wildfire-related PM2.5 concentration is joined to the health data. If FALSE,
the data set is loaded without any additional joins.}

\item{ncdf_path}{Path to a NetCDF file containing a daily time series of
gridded wildfire-related PM2.5 concentration data.}

\item{shp_path}{Path to a shapefile .shp of the geographical boundaries for
which to extract mean values of wildfire-related PM2.5}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names.}

\item{mean_temperature_col}{Character. Name of the column in the dataframe
that contains the mean temperature column.}

\item{health_outcome_col}{Character. Name of the column in the dataframe that
contains the health outcome count column (e.g. number of deaths, hospital
admissions)}

\item{pm_2_5_col}{Character. The name of the column containing PM2.5 values
in micrograms. Thsi is only required if health data isn't joined. Defaults to NULL.}
}
\value{
Dataframe containing a daily time series of
climate and health data.
}
\description{
Loads a dataframe containing a daily time series of health and
climate data, which may be disaggregated by region.
}
