% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_add_national_data}
\alias{mh_add_national_data}
\title{Process data for national analysis}
\usage{
mh_add_national_data(
  df_list,
  pop_list,
  var_fun = "bs",
  var_per = c(25, 50, 75),
  var_degree = 2,
  lag_fun = "strata",
  lag_breaks = 1,
  lag_days = 2,
  country = "National",
  cb_list,
  mm,
  minpercreg
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{pop_list}{List of population totals by year and region.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(25,50,75).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{lag_fun}{Character. Exposure function for arglag
(see dlnm::crossbasis). Defaults to 'strata'.}

\item{lag_breaks}{Integer. Internal cut-off point defining the strata for arglag
(see dlnm::crossbasis). Defaults to 1.}

\item{lag_days}{Integer. Maximum lag. Defaults to 2.
(see dlnm::crossbasis).}

\item{country}{Character. Name of country for national level estimates.}

\item{cb_list}{A list of cross-basis matrices by region.}

\item{mm}{A model object. A multivariate meta-analysis model.}

\item{minpercreg}{Vector. Percentile of minimum suicide temperature for each region.}
}
\value{
\itemize{
\item \code{df_list} List. A list of data frames for each region and nation.
\item \code{cb_list} List. A list of cross-basis matrices by region and nation.
\item \code{minpercreg} Vector. Percentile of minimum suicide temperature for each region and nation.
\item \code{mmpredall} List. A list of national coefficients and covariance matrices.
}
}
\description{
Aggregate to national data and run crossbasis
}
\keyword{internal}
