% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_attr_month}
\alias{mh_attr_month}
\title{Produce total attributable fractions, numbers and rates by calendar month}
\usage{
mh_attr_month(df_list, cb_list, minpercreg, maxpercreg, pred_list, pop_list)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}

\item{minpercreg}{Vector. Percentile of minimum suicide temperature for each area.}

\item{maxpercreg}{Vector. Percentile of maximum suicide temperature for each area.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{pop_list}{List of population totals by year and region.}
}
\value{
A list of data frames containing total attributable fractions, numbers and rates
by calendar month and area.
}
\description{
Create list of data frames containing total attributable fractions,
numbers and rates by calendar month over the whole time series.
}
