% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_attr_tables}
\alias{mh_attr_tables}
\title{Create attributable estimates tables}
\usage{
mh_attr_tables(attr_list, country = "National", meta_analysis = FALSE)
}
\arguments{
\item{attr_list}{A list containing attributable numbers per region.}

\item{country}{Character. Name of country for national level estimates.}

\item{meta_analysis}{Boolean. Whether to perform a meta-analysis.}
}
\value{
\itemize{
\item \code{res_attr_tot} Dataframe. Total attributable fractions, numbers and
rates for each area over the whole time series.
\item \code{attr_yr_list} List. Dataframes containing yearly estimates of
attributable fractions, numbers and rates by area.
\item \code{attr_mth_list} List. Dataframes containing total attributable
fractions, numbers and rates by calendar month and area.
}
}
\description{
Aggregate tables of attributable numbers, rates and fractions
for total, yearly and monthly by region and nation
}
