% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_attr_totals}
\alias{mh_attr_totals}
\title{Produce total attributable fractions, numbers and rates}
\usage{
mh_attr_totals(df_list, cb_list, minpercreg, maxpercreg, pred_list, pop_list)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}

\item{minpercreg}{Vector. Percentile of minimum suicide temperature for each area.}

\item{maxpercreg}{Vector. Percentile of maximum suicide temperature for each area.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{pop_list}{List of population totals by year and region.}
}
\value{
Matrix containing total attributable fractions, numbers and rates for each
area over the whole time series.
}
\description{
Create matrix of total attributable fractions, numbers and rates
for each area over the whole time series, with confidence intervals.
}
