% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_attr_yearly}
\alias{mh_attr_yearly}
\title{Produce yearly attributable fractions, numbers and rates}
\usage{
mh_attr_yearly(df_list, cb_list, minpercreg, maxpercreg, pred_list, pop_list)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}

\item{minpercreg}{Vector. Percentile of minimum suicide temperature for each area.}

\item{maxpercreg}{Vector. Percentile of maximum suicide temperature for each area.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{pop_list}{List of population totals by year and region.}
}
\value{
A list of matrices containing yearly estimates of attributable
fractions, numbers and rates by area
}
\description{
#' @description Create list of matrices of yearly attributable fractions,
numbers and rates for each area with confidence intervals.
}
