% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_casecrossover_dlnm}
\alias{mh_casecrossover_dlnm}
\title{Quasi-Poisson Case-Crossover model with DLNM}
\usage{
mh_casecrossover_dlnm(df_list, control_cols = NULL, cb_list)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{control_cols}{A list of confounders to include in the final model adjustment.
Defaults to NULL if none.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}
}
\value{
List containing models by region
}
\description{
Fits a quasi-Poisson case-crossover with a distributed lag
non-linear model
}
