% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_create_crossbasis}
\alias{mh_create_crossbasis}
\title{Create cross-basis matrix}
\usage{
mh_create_crossbasis(
  data,
  var_fun = "bs",
  var_degree = 2,
  var_per = c(25, 50, 75),
  lag_fun = "strata",
  lag_breaks = 1,
  lag_days = 2
)
}
\arguments{
\item{data}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(25,50,75).}

\item{lag_fun}{Character. Exposure function for arglag
(see dlnm::crossbasis). Defaults to 'strata'.}

\item{lag_breaks}{Integer. Internal cut-off point defining the strata for arglag
(see dlnm::crossbasis). Defaults to 1.}

\item{lag_days}{Integer. Maximum lag. Defaults to 2.
(see dlnm::crossbasis).}
}
\value{
A list of cross-basis matrices by region
}
\description{
Creates a cross-basis matrix for each region
}
