% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_meta_analysis}
\alias{mh_meta_analysis}
\title{Meta-analysis and BLUPs}
\usage{
mh_meta_analysis(data, coef_, vcov_)
}
\arguments{
\item{data}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{coef_}{A matrix of coefficients for the reduced model.}

\item{vcov_}{A list. Covariance matrices for each region for the reduced model.}
}
\value{
\itemize{
\item \code{mm} A model object. A multivariate meta-analysis model.
\item \code{blup} A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each region.
}
}
\description{
Run meta-analysis using temperature average and range as meta
predictors. Then create the best linear unbiased predictions (BLUPs).
}
