% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_min_suicide_temp}
\alias{mh_min_suicide_temp}
\title{Define minimum suicide temperature values}
\usage{
mh_min_suicide_temp(
  df_list,
  var_fun = "splines",
  var_per = c(25, 50, 75),
  var_degree = 2,
  blup = blup,
  coef_,
  meta_analysis = FALSE
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(25,50,75).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{blup}{A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each region.}

\item{coef_}{A matrix of coefficients for the reduced model.}

\item{meta_analysis}{Boolean. Whether to perform a meta-analysis.}
}
\value{
Vector. Percentile of minimum suicide temperature for each region.
}
\description{
Define the minimum suicide temperature values (between 1st and
50th percentiles).
}
