% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_minmax_suicide_temp}
\alias{mh_minmax_suicide_temp}
\title{Define min and max suicide values}
\usage{
mh_minmax_suicide_temp(
  data,
  var_fun = "bs",
  var_per = c(25, 50, 75),
  var_degree = 2,
  blup
)
}
\arguments{
\item{data}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(25,50,75).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{blup}{A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each region.}
}
\value{
\itemize{
\item \code{minpercreg} Vector. Percentile of minimum suicide temperature for each region.
\item \code{maxpercreg} Vector. Percentile of maximum suicide temperature for each region.
}
}
\description{
Define the minimum (between 1st and 50th percentiles) and
maximum (between 51st and 99th percentiles) suicide temperature values
}
