% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_model_validation}
\alias{mh_model_validation}
\title{Model Validation Assessment}
\usage{
mh_model_validation(
  df_list = df_list,
  cb_list = cb_list,
  independent_cols = NULL,
  save_fig = FALSE,
  save_csv = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}

\item{independent_cols}{Additional independent variables to test in model validation
as confounders.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{save_csv}{Boolean. Whether to save the results as a CSV. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}
}
\value{
\itemize{
\item \code{qaic_results} A dataframe of QAIC and dispersion metrics for each model combination and geography.
\item \code{qaic_summary} A dataframe with the mean QAIC and dispersion metrics for each model combination.
\item \code{vif_results} A dataframe. Variance inflation factors for each independent variables by region.
\item \code{vif_summary} A dataframe with the mean variance inflation factors for each independent variable.
}
}
\description{
Produces results on QAIC for each model combination, variance inflation
factor for each independent variable, and plots for residuals to assess the models
}
\keyword{internal}
