% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_plot_ar_monthly}
\alias{mh_plot_ar_monthly}
\title{Plot attributable rates by calendar month}
\usage{
mh_plot_ar_monthly(
  attr_mth_list,
  df_list,
  country = "National",
  save_fig = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{attr_mth_list}{A list of data frames containing total attributable
fractions, numbers and rates by calendar month and area.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}
}
\value{
Plots of attributable rates by calendar month per area
}
\description{
Plot attributable rates grouped over the whole time series by
calendar month to explore seasonality.
}
