% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_plot_power}
\alias{mh_plot_power}
\title{Plot power}
\usage{
mh_plot_power(
  power_list,
  save_fig = FALSE,
  output_folder_path = NULL,
  country = "National"
)
}
\arguments{
\item{power_list}{A list containing power information by area.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}

\item{country}{Character. Name of country for national level estimates.
Defaults to 'National'}
}
\value{
Plots of power by temperature for the attributable threshold and above
for each area.
}
\description{
Plots the power statistic for each reference temperature at and above
the attributable risk threshold for each area.
}
\keyword{internal}
