% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_plot_results}
\alias{mh_plot_results}
\title{Plot results of analysis - Mental Health}
\usage{
mh_plot_results(pred_list, save_fig = FALSE, output_folder_path = NULL)
}
\arguments{
\item{pred_list}{A list containing predictions from the model by region.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}
}
\value{
Plots of cumulative lag exposure-response function for each region
}
\description{
Plots cumulative lag exposure-response function for each region
}
