% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_plot_rr}
\alias{mh_plot_rr}
\title{Plot results of relative risk analysis - Mental Health}
\usage{
mh_plot_rr(
  df_list,
  pred_list,
  attr_thr = 97.5,
  minpercreg,
  country = "National",
  save_fig = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{attr_thr}{Integer. Percentile at which to define the temperature threshold for
calculating attributable risk.}

\item{minpercreg}{Vector. Percentile of minimum suicide temperature for each area.}

\item{country}{Character. Name of country for national level estimates.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}
}
\value{
Plots of cumulative lag exposure-response function with histogram of
temperature distribution for each region
}
\description{
Plots cumulative lag exposure-response function with histogram of
temperature distribution for each region
}
\keyword{internal}
