% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_pop_totals}
\alias{mh_pop_totals}
\title{Create population totals}
\usage{
mh_pop_totals(df_list, country = "National")
}
\arguments{
\item{country}{Character. Name of country for national level estimates.}

\item{data}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}
}
\value{
List of population totals by year and region
}
\description{
Creates a list of population totals by year and region for use
in the attributable rate calculations.
}
