% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_power_list}
\alias{mh_power_list}
\title{Mental Health Power calculation}
\usage{
mh_power_list(df_list, pred_list, minpercreg, attr_thr = 97.5)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{minpercreg}{Vector. Percentile of maximum suicide temperature for each region.}

\item{attr_thr}{Integer. Percentile at which to define the temperature threshold for
calculating attributable risk. Defaults to 97.5.}
}
\value{
A list containing power information by area.
}
\description{
Produce a power statistic by area for the attributable threshold
and above as a reference.
}
\keyword{internal}
