% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_predict}
\alias{mh_predict}
\title{Run predictions from model}
\usage{
mh_predict(data, cb_list, model_list)
}
\arguments{
\item{data}{Dataframe containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}

\item{model_list}{List of models produced from case-crossover and DLNM
analysis.}
}
\value{
A list containing predictions by region
}
\description{
Use model to run predictions
}
