% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_read_and_format_data}
\alias{mh_read_and_format_data}
\title{Read in and format data - Mental Health}
\usage{
mh_read_and_format_data(
  data_path,
  date_col,
  region_col = NULL,
  temperature_col,
  health_outcome_col,
  population_col
)
}
\arguments{
\item{data_path}{Path to a csv file containing a daily time series of data
for a particular health outcome and climate variables, which may be
disaggregated by region.}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names. Defaults to NULL.}

\item{temperature_col}{Character. Name of the column in the dataframe that
contains the temperature column.}

\item{health_outcome_col}{Character. Name of the column in the dataframe that
contains the health outcome count column (e.g. number of deaths, hospital
admissions).}

\item{population_col}{Character. Name of the column in the dataframe that
contains the population estimate coloumn.}
}
\value{
A list of dataframes with formatted and renamed columns.
}
\description{
Reads in a CSV file for a daily time series of health and climate
data, renames columns and creates stratum for case-crossover analysis.
}
\keyword{internal}
