% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_rr_results}
\alias{mh_rr_results}
\title{Produce cumulative relative risk results of analysis}
\usage{
mh_rr_results(pred_list)
}
\arguments{
\item{pred_list}{A list containing predictions from the model by region.}
}
\value{
Dataframe containing cumulative relative risk and confidence
intervals from analysis.
}
\description{
Produces cumulative relative risk and confidence intervals
from analysis.
}
