% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{ncdf_to_shp}
\alias{ncdf_to_shp}
\title{Extract mean values for shapefile regions from NetCDF file}
\usage{
ncdf_to_shp(ncdf_path, shapefile_path)
}
\arguments{
\item{ncdf_path}{Path to a NetCDF file containing a daily time series of
gridded wildfire-related PM2.5 concentration data. We use open-source data
from the Finnish Meteorological Institute's SILAM-CTM model
https://doi.org/10.57707/fmi-b2share.d1cac971b3224d438d5304e945e9f16c}

\item{shp_path}{Path to a shapefile .shp of the geographical boundaries for
which to extract mean values of wildfire-related PM2.5}
}
\value{
Dataframe containing a daily time series with mean wildfire-related
PM2.5 values for each region
}
\description{
Takes a NetCDF file of gridded wildfire data and shapefile for
geographical regions and extracts mean values for each shapefile region.
}
