% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{pair_with_health}
\alias{pair_with_health}
\title{Join health and climate data}
\usage{
pair_with_health(
  climate_data,
  health_data,
  date_col,
  region_col = NULL,
  mean_temperature_col,
  health_outcome_col
)
}
\arguments{
\item{climate_data}{Dataframe containing a daily time series of climate data,
which may be disaggregated by region.}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names.}

\item{mean_temperature_col}{Character. Name of the column in the dataframe
that contains the mean temperature column.}

\item{health_outcome_col}{Character. Name of the column in the dataframe that
contains the health outcome count column (e.g. number of deaths, hospital
admissions)}

\item{health_path}{Path to a CSV file containing a
daily time series of data for a particular health outcome, which may be
disaggregated by region.}
}
\value{
Dataframe containing a daily time series of the joined
climate and health data.
}
\description{
Joins a daily time series of climate data with a daily time
series of health data.
}
