% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{plot_aggregated_AF}
\alias{plot_aggregated_AF}
\title{Plots attributable fractions and CI across years by regions}
\usage{
plot_aggregated_AF(
  data,
  by_region = FALSE,
  include_all_regions = TRUE,
  output_dir = "."
)
}
\arguments{
\item{data}{A data frame containing annual attributable fraction estimates.
Must include columns: year, average_attributable_fraction,
lower_ci_attributable_fraction, upper_ci_attributable_fraction. If
by_region is TRUE, must also include region.}

\item{by_region}{Logical. If TRUE, plots are generated per region using
region. Defaults to FALSE.}

\item{include_all_regions}{Logical. If TRUE includes All regions plots in output dir}

\item{output_dir}{Character. Directory path where the PDF file will be
saved. Must exist. Defaults to ".".}
}
\value{
No return value. A PDF file is created.
}
\description{
Generates a PDF containing one or more plots of average attributable
fractions over time. If by_region is TRUE, the function creates separate
plots for each region. All plots are saved to a single PDF
file named "aggregated_AF_by_region.pdf" in the specified output_dir.
}
\keyword{internal}
