% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{plot_air_pollution_dlm}
\alias{plot_air_pollution_dlm}
\title{Create distributed lag plots for countries (national) and regions}
\usage{
plot_air_pollution_dlm(
  dlm_results,
  output_dir = NULL,
  max_lag = 2,
  width_per_panel = 4,
  height_per_panel = 4
)
}
\arguments{
\item{dlm_results}{Results from climatehealth::analyze_air_pollution_dlm.}

\item{output_dir}{String. Directory to save plots. Defaults to NULL.}

\item{max_lag}{Integer. Maximum lag days. Defaults to to 2.}

\item{width_per_panel}{Integer. Width per panel for saving. Defaults to 4.}

\item{height_per_panel}{Integer. Height per panel for saving. Defaults to 4.}
}
\value{
List of ggplot objects.
}
\description{
Create distributed lag plots for countries (national) and regions.
}
\keyword{internal}
