% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{plot_air_pollution_power}
\alias{plot_air_pollution_power}
\title{Plot Power vs PM2.5 Concentration}
\usage{
plot_air_pollution_power(
  power_list,
  output_dir = NULL,
  save_plot = FALSE,
  ref_name = "WHO",
  include_national = TRUE
)
}
\arguments{
\item{power_list}{A list containing power information by region.}

\item{output_dir}{Character. Directory to save plot. Defaults to NULL.}

\item{save_plot}{Logical. Whether to save the plot. Defaults to FALSE.}

\item{ref_name}{Character. Reference standard name for plot title.}

\item{include_national}{Logical. Whether to include national level in the plot.
Defaults to TRUE.}
}
\value{
Invisible list of plot information
}
\description{
Plots the power statistic for each reference PM2.5 at and above
the attributable risk threshold for each region.
}
\keyword{internal}
