% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{plot_air_pollution_variables}
\alias{plot_air_pollution_variables}
\title{Create a scatterplot}
\usage{
plot_air_pollution_variables(
  data,
  xvar,
  yvar,
  output_dir = NULL,
  save_plot = FALSE
)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate, environmental
and health data}

\item{xvar}{Character. x variable}

\item{yvar}{Character. y variable}

\item{output_dir}{Character. Directory to save plot. Defaults NULL.}

\item{save_plot}{Logical. Whether to save the plot. Defaults FALSE.}
}
\description{
Produces a ggplot2 scatterplot of two variables x versus y.
}
