% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{plot_and_write}
\alias{plot_and_write}
\title{Plot and write results of analysis}
\usage{
plot_and_write(
  df_list,
  output_name,
  aggregate_outputs = FALSE,
  output_folder_path = "",
  save_fig = TRUE,
  save_csv = TRUE,
  cb = NULL,
  model = NULL,
  blup = NULL,
  mintempregions,
  an_thresholds,
  independent_cols = NULL,
  varfun,
  varper,
  vardegree,
  lag = NULL,
  lagnk = NULL,
  dfseas = NULL,
  dependent_col = NULL
)
}
\arguments{
\item{df_list}{An alphabetically-ordered
list of dataframes for each region.}

\item{output_name}{The name of the output file. (.csv and .pdf added
accordingly).}

\item{aggregate_outputs}{Whether or not to output all geographical regions.}

\item{output_folder_path}{The directory to output the resultant data/plots to.}

\item{save_fig}{Whether to save output figure (Bool)}

\item{save_csv}{Whether to save output CSVs (Bool)}

\item{cb}{The generated crossbasis.}

\item{model}{The generated model.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{mintempregions}{A named numeric vector.
Minimum (optimum) mortality temperature per region.}

\item{an_thresholds}{A dataframe with the optimal temperature range and
temperature thresholds for calculation of attributable deaths.}

\item{independent_cols}{column name (or list of names) of extra independent
variable to include in regression (excluding temperature). Defaults to NULL.}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}

\item{dependent_col}{the column name of the
dependent variable of interest e.g. deaths}
}
\value{
\itemize{

\item A PDF containing a line plot of temperature versus relative risk per
region, and histogram of temperatures per region.
\item A CSV of relative risk per temperature per region.
\item \code{output_df} A dataframe with relative risk estimates and confidence
intervals across the temperature range for each region.
\item \code{temp_df} A dataframe with daily mean exposure values for each
region.
}
}
\description{
Plot and write results of analysis
}
