% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{plot_and_write_relative_risk}
\alias{plot_and_write_relative_risk}
\title{Plot and write results of analysis}
\usage{
plot_and_write_relative_risk(
  df_list,
  blup = NULL,
  mintempregions,
  an_thresholds,
  save_fig = TRUE,
  save_csv = TRUE,
  csv_output_path = NULL,
  independent_cols,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas
)
}
\arguments{
\item{df_list}{An alphabetically-ordered
list of dataframes for each region.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{mintempregions}{A named numeric vector.
Minimum (optimum) mortality temperature per region.}

\item{an_thresholds}{A dataframe with the optimal temperature range and
temperature thresholds for calculation of attributable deaths.}

\item{save_fig}{Whether to save output figure (Bool)}

\item{save_csv}{Whether to save output CSVs (Bool)}

\item{csv_output_path}{The path to save the csv to, including file name and
file extensions}

\item{independent_cols}{column name (or list of names) of extra independent
variable to include in regression (excluding temperature). Defaults to NULL.}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}
}
\value{
\itemize{

\item A PDF containing a line plot of temperature versus relative risk per
region, and histogram of temperatures per region.
\item A CSV of relative risk per temperature per region.
\item \code{output_df} A dataframe with relative risk estimates and confidence
intervals across the temperature range for each region.
\item \code{temp_df} A dataframe with daily mean exposure values for each
region.
}
}
\description{
Plot and write results of analysis
}
\examples{
csv_output_path = "directory/sub_directory/file_name.csv"

}
