% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{plot_and_write_relative_risk_all}
\alias{plot_and_write_relative_risk_all}
\title{Plot and write results of analysis}
\usage{
plot_and_write_relative_risk_all(
  df_list,
  mintempregions,
  save_fig = TRUE,
  save_csv = TRUE,
  dependent_col,
  varfun,
  varper,
  vardegree,
  coef,
  vcov,
  output_folder_path
)
}
\arguments{
\item{df_list}{An alphabetically-ordered
list of dataframes for each region.}

\item{mintempregions}{A named numeric vector.
Minimum (optimum) mortality temperature per region.}

\item{save_fig}{Whether to save output figure (Bool)}

\item{save_csv}{Whether to save output CSVs (Bool)}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{output_folder_path}{Path to folder for storing outputs.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{regions_df}{A dataframe with two columns.
Column 1 is abbreviated region names.
Column 2 is user-specified region names.}

\item{indepedent_col}{column names of independent
variables to include in regression (excluding temperature,
see config file for formula structure)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}
}
\value{
A PDF containing a line plot of temperature
versus relative risk per region,
and histogram of temperatures per region.
A CSV of relative risk per temperature per region.
}
\description{
Plot and write results of analysis
}
\examples{
output_folder_path = 'myfolder/output/'
}
