% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{plot_ar_pm_monthly}
\alias{plot_ar_pm_monthly}
\title{Plot monthly deaths and PM2.5 concentrations with dual y-axes}
\usage{
plot_ar_pm_monthly(
  data,
  save_outputs = FALSE,
  output_dir = NULL,
  include_all_regions = TRUE
)
}
\arguments{
\item{data}{A data frame with columns: month, deaths_per_100k, and
monthly_avg_pm25. Month names must match month.abb.}

\item{save_outputs}{Logical. If TRUE, saves the plot as PNG and the
aggregated data as CSV. Defaults to FALSE.}

\item{output_dir}{Character. Directory path where outputs are saved if
save_outputs is TRUE. Must exist. Defaults to NULL.}

\item{include_all_regions}{Logical. If TRUE includes All regions plots in output dir}
}
\value{
No return value. Generates a plot and optionally saves files.
}
\description{
Aggregates data by month and creates a dual-axis plot showing average
deaths per 100,000 and mean PM2.5 concentrations.
}
\keyword{internal}
