% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Plot a correlation matrix as an accessible heatmap.}
\usage{
plot_correlation_matrix(matrix_, title, output_path)
}
\arguments{
\item{matrix_}{The matrix to plot.}

\item{title}{The title for the correlation matrix.}

\item{output_path}{The path to output the plot to.}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Plot a correlation matrix as an accessible heatmap.
}
\keyword{internal}
