% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_rate_overall}
\alias{plot_rate_overall}
\title{Plot the rate of a dependent variable per 100,000 population per year.}
\usage{
plot_rate_overall(
  df,
  dependent_col,
  population_col,
  date_col,
  save_rate = FALSE,
  output_path = NULL
)
}
\arguments{
\item{df}{The dataframe containing the data.}

\item{dependent_col}{The name of the column representing the dependent variable.}

\item{population_col}{The name of the column representing the population.}

\item{date_col}{The name of the column containing date values.}

\item{save_rate}{Whether to save the plot as a PDF.}

\item{output_path}{The file path to save the plot if save_rate is TRUE.}
}
\description{
Plot the rate of a dependent variable per 100,000 population per year.
}
\keyword{internal}
