% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_regional_trends}
\alias{plot_regional_trends}
\title{Plot regional trends of a climate and healthoutcome.}
\usage{
plot_regional_trends(
  df,
  region_col,
  outcome_cols,
  title = "Regional Trends",
  ylabs = NULL,
  save_plot = FALSE,
  output_path = ""
)
}
\arguments{
\item{df}{The dataframe containing the raw data.}

\item{region_col}{The name of the column containing regions.}

\item{title}{The title of your plot.}

\item{ylabs}{A character vector of y-axis labels (e.g., with units) corresponding to the columns.}

\item{save_plot}{Whether or not to save the plot.}

\item{output_path}{The path to output the plot to.}

\item{outcome_col}{The name of the outcome column to analyse.}
}
\description{
Plot regional trends of a climate and healthoutcome.
}
