% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{plot_relative_risk}
\alias{plot_relative_risk}
\title{Read in Relative Risk plot at country, Region, and District level}
\usage{
plot_relative_risk(
  data,
  model,
  param_term,
  level = "country",
  filter_year = NULL,
  case_type,
  output_dir = NULL,
  save_csv = FALSE,
  save_fig = FALSE
)
}
\arguments{
\item{data}{Data list from combine_health_climate_data() function.}

\item{model}{The fitted model from run_inla_models() function.}

\item{param_term}{A character vector or list containing parameter terms such
as \code{tmax} (temperature exposure) and \code{rainfall} (rainfall exposure).
Default to \code{tmax}.}

\item{level}{A character vector specifying the geographical disaggregation.
Can take one of the following values: \code{country}, \code{region}, or \code{district}.
Default to \code{country}.}

\item{filter_year}{Integer. The year to filter to data to. This gives the
possibility to user to have the plot for a specific year. When Defaults to NULL,
it provides the plot by grouping all the years in the dataset.}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of \code{"diarrhea"} or \code{"malaria"}.}

\item{output_dir}{Character. The path where the PDF file will be saved. Default to NULL.}

\item{save_csv}{Boolean. If TRUE, saves the RR data to the specified directory.
Defaults to FALSE.}

\item{save_fig}{Boolean. If TRUE, saves the plot to the specified directory.
Defaults to FALSE.}
}
\value{
Relative risk plot at country, region, and district levels.
}
\description{
Plots the relative risk of Malaria cases by the maximum
temperature and cumulative rainfall at country, Region and District level
}
\keyword{internal}
