% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{plot_results}
\alias{plot_results}
\title{Plot results of analysis}
\usage{
plot_results(results, save_fig, wildfire_lag = 3, output_folder_path)
}
\arguments{
\item{results}{Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5}

\item{save_fig}{Boolean. Whether to save the plot as an output.}

\item{wildfire_lag}{Integer. The maximum number of days for which to plot the
lags for wildfire PM2.5. Default is 3.}

\item{output_folder_path}{Path to folder where plots should be saved.}
}
\value{
Plot of relative risk and confidence intervals for each lag of
wildfire-related PM2.5
}
\description{
Plots relative risk and confidence intervals for each lag value
of wildfire-related PM2.5
}
