% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{plot_RR}
\alias{plot_RR}
\title{Plot relative risk results by region.}
\usage{
plot_RR(
  rr_data,
  wildfire_lag = 3,
  by_region = FALSE,
  save_fig = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{rr_data}{Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5}

\item{wildfire_lag}{Integer. The maximum number of days for which to plot the
lags for wildfire PM2.5. Defaults to 3.}

\item{by_region}{Bool. Whether to plot RR(relative risk) by region.
Defaults to FALSE}

\item{save_fig}{Boolean. Whether to save the generated plot.
Defaults to FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.}
}
\value{
Plot of relative risk and confidence intervals for each lag of
wildfire-related PM2.5
}
\description{
Plots relative risk and confidence intervals for each lag value
of wildfire-related PM2.5.
}
\keyword{internal}
