% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{plot_RR_by_region}
\alias{plot_RR_by_region}
\title{Plot relative risk results.}
\usage{
plot_RR_by_region(
  results,
  save_fig = FALSE,
  wildfire_lag = 3,
  relative_risk_by_region = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{results}{Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to FALSE.}

\item{wildfire_lag}{Integer. The maximum number of days for which to plot the
lags for wildfire PM2.5. Defaults to 3.}

\item{relative_risk_by_region}{Bool. Whether to calculate Relative Risk by region.
Defaults to FALSE}

\item{output_folder_path}{Path to folder where plots should be saved.}
}
\value{
Plot of relative risk and confidence intervals for each lag of
wildfire-related PM2.5
}
\description{
Plots relative risk and confidence intervals for each lag value
of wildfire-related PM2.5. If RR by region is true, plots RR results by region.
If false plots overall RR.
}
\keyword{internal}
