% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{plot_RR_core}
\alias{plot_RR_core}
\title{Core functionality for plotting results of relative risk analysis.}
\usage{
plot_RR_core(
  rr_data,
  save_fig = FALSE,
  wildfire_lag = 3,
  output_folder_path = NULL,
  region_name = "All regions",
  ylims = NULL
)
}
\arguments{
\item{rr_data}{Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5.}

\item{save_fig}{Boolean. Whether to save the plot as an output.
Defaults to FALSE.}

\item{wildfire_lag}{Integer. The maximum number of days for which to plot the
lags for wildfire PM2.5. Defaults to 3.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}

\item{region_name}{Character. The name of the region.
Defaults to 'All regions'.}
}
\value{
Plot of relative risk and confidence intervals for each lag of
wildfire-related PM2.5.
}
\description{
Plots relative risk and confidence intervals for each lag value
of wildfire-related PM2.5.
}
\keyword{internal}
