% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_scatter_grid}
\alias{plot_scatter_grid}
\title{Plot a grid of scatter graphs comparing one column to various others.}
\usage{
plot_scatter_grid(
  df,
  main_col,
  comparison_cols,
  title,
  save_scatters = FALSE,
  output_path = "",
  units = NULL
)
}
\arguments{
\item{df}{The dataframe containing the raw data.}

\item{main_col}{The main column to compare with all other columns.}

\item{comparison_cols}{The columns to compare with.}

\item{title}{The title of your plot.}

\item{save_scatters}{Whether or not to save the plot.}

\item{output_path}{The path to output the plot to.}

\item{units}{A named character vector of units for each variable.}
}
\description{
Plot a grid of scatter graphs comparing one column to various others.
}
