% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_total_variables_by_year}
\alias{plot_total_variables_by_year}
\title{Plot the total of selected variables per year.}
\usage{
plot_total_variables_by_year(
  df,
  date_col,
  variables,
  title,
  save_total = FALSE,
  output_path = ""
)
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{date_col}{The name of the column containing date values.}

\item{variables}{Column names to be summed and plotted.}

\item{title}{Character. The specific title for the subset of data being used.}

\item{save_total}{if TRUE, saves each plot as a PDF.}

\item{output_path}{The file path for saving plots.}
}
\value{
Plots are printed to the console or saved as PDF files.
}
\description{
Plot the total of selected variables per year.
}
\keyword{internal}
