% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{plot_variables}
\alias{plot_variables}
\title{Scatterplot}
\usage{
plot_variables(data, xvar, yvar)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data}

\item{xvar}{x variable}

\item{yvar}{y variable}
}
\value{
Prints a ggplot2 scatterplot of x versus y
}
\description{
Produces a ggplot2 scatterplot of two variables x versus y.
}
\keyword{internal}
