% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{prep_and_first_step}
\alias{prep_and_first_step}
\title{First-stage analysis
Run the model in each city/region, reduce and save.
Create objects to store the results.}
\usage{
prep_and_first_step(input_csv_path)
}
\arguments{
\item{input_csv_path}{A character.
The file path for a csv file of the input data.}
}
\value{
A list of the following: \cr
dlist: a list of dataframes for each region \cr
argvar: A list of arguments to pass to \link[dlnm]{crossbasis} function \cr
regions: A character vector.
Contains names of ten regions; the 9 English regions and Wales. \cr
cities: A dataframe with two columns.
Column 1 is abbreviated region names.
Column 2 is full region names \cr
coef: A matrix populated with predicted values of ??? from a model of ??? \cr
vcov: A list ??? \cr
}
\description{
First-stage analysis
Run the model in each city/region, reduce and save.
Create objects to store the results.
}
\examples{
prep_and_first_step('data/regEngWales.csv')
}
