% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{prepare_descriptive_input}
\alias{prepare_descriptive_input}
\title{Normalise descriptive stats data input to combined and regional dataframes.}
\usage{
prepare_descriptive_input(
  data,
  aggregation_column = NULL,
  timeseries_col = NULL
)
}
\arguments{
\item{data}{Dataframe or list of dataframes.}

\item{aggregation_column}{Character. Region column for splitting dataframes.}

\item{timeseries_col}{Character. Date column for timeseries analysis.}
}
\value{
A list with \code{combined_df} and \code{region_df_list}.
}
\description{
Normalise descriptive stats data input to combined and regional dataframes.
}
\keyword{internal}
