% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{produce_results}
\alias{produce_results}
\title{Produce cumulative relative risk results of analysis}
\usage{
produce_results(data, cb_list, model_list)
}
\arguments{
\item{data}{Dataframe containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}

\item{model_list}{List of models produced from case-crossover and DLNM
analysis.}
}
\value{
Dataframe containing cumulative relative risk and confidence
intervals from analysis.
}
\description{
Produces cumulative relative risk and confidence intervals
from analysis.
}
