% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{raise_if_null}
\alias{raise_if_null}
\title{Raise an Error if a Parameter's Value is NULL}
\usage{
raise_if_null(param_nm, value)
}
\arguments{
\item{param_nm}{Character. The parameter name.}

\item{value}{Any. The value of the parameter.}
}
\value{
None. Stops execution if value is NULL.
}
\description{
Raise an Error if a Parameter's Value is NULL
}
\keyword{internal}
