% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{read_input_data}
\alias{read_input_data}
\title{Read a csv file into memory as a data frame.}
\usage{
read_input_data(input_csv_path)
}
\arguments{
\item{input_csv_path}{The path to the csv to read as a dataframe.}
}
\value{
A dataframe containing the data from the csv.
}
\description{
Read a csv file into memory as a data frame.
}
\examples{
input_csv_path = "directory/file_name.csv"

}
\keyword{internal}
