% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_utils.R
\name{reformat_data}
\alias{reformat_data}
\title{Reformat a dataframe using various different cleaning techniques.}
\usage{
reformat_data(df, reformat_date = TRUE, fill_na = c(), year_from_date = TRUE)
}
\arguments{
\item{df}{The dataframe to apply cleaning/reformatting to.}

\item{reformat_date}{Whether or not to reformat the data to the Date datatype.}

\item{fill_na}{A vector of column names to fill NA values in (fills with 0).}

\item{year_from_date}{Derive a new column 'year' from the date column.}
}
\value{
The cleaned/reformatted data frame.
}
\description{
Take a dataframe, and apply various different cleaning methods to it in order
to prepare the data for use with a climate indicator.
}
