% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{relative_risk_by_region}
\alias{relative_risk_by_region}
\title{Passes data to casecrossover_quasipoisson to calculate RR.}
\usage{
relative_risk_by_region(
  data,
  scale_factor = 10,
  wildfire_lag = 3,
  calc_relative_risk_by_region = FALSE,
  save_fig = FALSE,
  output_folder_path = NULL,
  print_model_summaries = FALSE
)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data from which to fit models.}

\item{scale_factor}{Numeric. The value to divide the wildfire PM2.5
concentration variables by for alternative interpretation of outputs.
Corresponds to the unit increase in wildfire PM2.5 to give the model
estimates and relative risks (e.g. scale_factor = 10 corresponds to estimates
and relative risks representing impacts of a 10 unit increase in wildfire PM2.5)
Setting this parameter to 0 or 1 leaves the variable unscaled.}

\item{wildfire_lag}{Integer. The maximum number of days for which to calculate
lagged results for wildfire PM2.5. Default is 3.}

\item{calc_relative_risk_by_region}{Bool. Whether to calculate Relative Risk by region.
Default: FALSE}

\item{save_fig}{Bool. Whether or not to save a figure showing residuals vs
fitted values for each lag. Defaults to FALSE.}

\item{output_folder_path}{String. Where to save the figure. Defaults to NULL.}

\item{print_model_summaries}{Bool. Whether to print the model summaries to
console. Defaults to FALSE.}
}
\value{
Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5. Split by region if calc_relative_risk_by_region
set to TRUE.
}
\description{
Splits data by region if relative_risk_by_region config option is
TRUE. If true data for each individual region is passed to casecrossover_quasipoisson
to calculate RR by region. If false RR is calculated for the entire dataset.
}
