% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{resolve_input_column_name}
\alias{resolve_input_column_name}
\title{Resolve input column name}
\usage{
resolve_input_column_name(
  column,
  available,
  dataset_name = "dataset",
  argument_name = "column",
  call = rlang::caller_env()
)
}
\arguments{
\item{column}{Character. Imput column name}

\item{available}{Character. vector of available column names}

\item{dataset_name}{Character. Name of data frame containing columns}

\item{argument_name}{Character. Name of the input argument.}
}
\value{
Character. Resolved column name
}
\description{
resolves and validates column reference supplied by the user against
available columns in the input dataset
}
\keyword{internal}
