% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{save_results}
\alias{save_results}
\title{Save results of analysis}
\usage{
save_results(results, output_folder_path)
}
\arguments{
\item{results}{Dataframe containing cumulative relative risk and confidence
intervals from analysis.}

\item{output_folder_path}{Path to folder where results should be saved.}
}
\description{
Saves a CSV file of cumulative relative risk and
confidence intervals.
}
