% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{second_stage}
\alias{second_stage}
\title{Second-stage analysis
Multivariate meta-analysis of the reduced
coef and then computation of blup}
\usage{
second_stage(dlist, cities, argvar, coef, vcov)
}
\arguments{
\item{vcov}{}
}
\value{
A list of the following: \cr
}
\description{
Second-stage analysis
Multivariate meta-analysis of the reduced
coef and then computation of blup
}
\examples{
second_stage(dlist = dlist, cities = cities, argvar = argvar,
coef = coef, vcov = vcov)
}
