% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{set_cross_basis}
\alias{set_cross_basis}
\title{Create a cross-basis matrix set for DLNM analysis}
\usage{
set_cross_basis(data, max_lag = 2, nk = 2)
}
\arguments{
\item{data}{A dataset returned from \code{combine_health_climate_data()},
including lagged variables like \code{tmax_lag1}, \code{tmin_lag1}, etc.}

\item{max_lag}{Character. Number corresponding to the maximum lag to be
considered for the delay effect. It should be between 2 an 4. Defaults to 2.}

\item{nk}{Numeric. Number of internal knots for the natural spline of
each predictor, controlling its flexibility: \code{nk = 0} produces a linear
effect with one basis column, \code{nk = 1} generates a simple spline with two
columns, \code{nk = 2} yields a more flexible spline with three columns,
and higher values of \code{nk} further increase flexibility but may also
raise collinearity among spline terms. Defaults to 1}
}
\value{
A list of cross-basis matrices including the basis matrix for maximum
temperature, minimun temperature, cumulative rainfall, and relative humidity.
}
\description{
Generates cross-basis matrices for lagged climate variables in a dataset,
for use in Distributed Lag Nonlinear Models (DLNM).
}
\keyword{internal}
