% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{third_stage}
\alias{third_stage}
\title{Third-stage analysis
Compute the attributable deaths for each city,
with emprical CI estimated using the re-centred bases}
\usage{
third_stage(
  dlist,
  regions,
  cities,
  coef,
  vcov,
  varfun,
  argvar,
  bvar,
  blup,
  mintempcity
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A number.
}
\description{
Third-stage analysis
Compute the attributable deaths for each city,
with emprical CI estimated using the re-centred bases
}
