% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{time_stratify}
\alias{time_stratify}
\title{Stratify data by time period}
\usage{
time_stratify(data)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data}
}
\value{
Dataframe with additional columns for stratum
(region:year:month:dayofweek) and for the total counts of a health outcome
across days in each stratum.
}
\description{
Adds columns for strata for each region:year:month:dayofweek
and for the total counts of a health outcome across days in each stratum.
}
