% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{validate_case_type}
\alias{validate_case_type}
\title{Ensure that the \code{case_type} parameter is valid}
\usage{
validate_case_type(case_type)
}
\arguments{
\item{case_type}{Character. The value of the case_type parameter.}
}
\value{
Character. The lower case_type.
}
\description{
Ensures that the case_type parameter is either malaria or
diarrhea to comply with supported indicators.
}
\keyword{internal}
