% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{wald_results}
\alias{wald_results}
\title{Get Wald statistic for a meta-analysis model}
\usage{
wald_results(mv)
}
\arguments{
\item{mv}{A model object (multivariate meta-analysis model)}
}
\value{
P-values for average and range of temperatures
(avgtmean_wald, rangetmean_wald).
}
\description{
Get Wald statistic for a meta-analysis model
}
