% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{write_attributable_deaths}
\alias{write_attributable_deaths}
\title{Write outputs to csv}
\usage{
write_attributable_deaths(
  avgtmean_wald,
  rangetmean_wald,
  anregions_bind,
  antot_bind,
  arregions_bind,
  artot_bind,
  output_folder_path = NULL
)
}
\arguments{
\item{anregions_bind}{A dataframe with attributable deaths for each region.}

\item{arregions_bind}{A dataframe with attributable rates by region.}

\item{output_folder_path}{Path to folder for storing outputs.}

\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{totdeath}{A named vector of integers. Total observed mortality per region.}

\item{arraysim}{An array (numeric). Total (glob),
cold and heat-attributable deaths per region for 1000 simulations.
Used to derive confidence intervals.}

\item{matsim}{A matrix (numeric). Total (glob),
cold and heat-attributable deaths per region from reduced coefficients.}
}
\value{
\itemize{
\item \code{anregions_bind} A matrix of numbers of deaths attributable to
temperature, heat, cold, extreme heat and extreme cold (with confidence
intervals), disaggregated by region.
\item \code{antot_bind} A matrix of numbers of numbers of deaths attributable to
temperature, heat, cold, extreme heat and extreme cold (with confidence
intervals).
\item \code{arregions_bind} A matrix of fractions of all-cause mortality
attributable to temperature, heat, cold, extreme heat and extreme cold
(with confidence intervals), disaggregated by region.
\item \code{artot_bind} A matrix of fractions of all-cause mortality
attributable to temperature, heat, cold, extreme heat and extreme cold
(with confidence intervals).
}
}
\description{
Write the attributable deaths and temperature for each regions,
with empirical CI estimated using the re-centered bases.
}
\examples{
output_folder_path = 'myfolder/output/'
}
