% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{write_outputs_to_csv}
\alias{write_outputs_to_csv}
\title{Write outputs to csv
Write the attributable deaths and temperature for each regions,
with empirical CI estimated using the re-centred bases}
\usage{
write_outputs_to_csv(
  regions_df,
  matsim,
  arraysim,
  totdeath,
  output_folder_path = NULL
)
}
\arguments{
\item{output_folder_path}{}
}
\value{
None
}
\description{
Write outputs to csv
Write the attributable deaths and temperature for each regions,
with empirical CI estimated using the re-centred bases
}
\examples{
output_folder_path = 'myfolder/output/'
}
