% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{air_pollution_meta_analysis}
\alias{air_pollution_meta_analysis}
\title{Perform meta analysis with multiple lag structures}
\usage{
air_pollution_meta_analysis(
  data_with_lags,
  max_lag = 14L,
  df_seasonal = 6L,
  family = "quasipoisson"
)
}
\arguments{
\item{data_with_lags}{Lagged data}

\item{max_lag}{Integer. Maximum lag days. Defaults to 14}

\item{df_seasonal}{Integer. Degrees of freedom for seasonal spline. Default 6.}

\item{family}{Character string indicating the distribution family used in the GAM.}
}
\value{
Dataframe with lag-specific results including for regional and national
}
\description{
Implements distributed lag model. Individual lag
coefficients and cumulative effects are extracted and perform meta analysis
}
\keyword{internal}
