% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{air_pollution_meta_analysis}
\alias{air_pollution_meta_analysis}
\title{Conduct two-stage meta-analysis with random effect with AF and AN calculations}
\usage{
air_pollution_meta_analysis(data, var_name = "pm25", family = "quasipoisson")
}
\arguments{
\item{data}{Input data with multiple regions}

\item{var_name}{Character. PM2.5 variable to analyze. Defaults to "pm25".}

\item{family}{Character string indicating the distribution family used in the GAM.
Must be one of the four options "quasipoisson", "nb", "ziP", and "poisson".
"quasipoisson" (Appropriate for count data with overdispersion),
"nb" (negativebinomial which is suitable for highly dispersed count data),
"ziP" (zero-inflated which is suitable for count data with excess zeros (sparse data)),
"poisson" (Suitable for count data without overdispersion (mean = variance)).
Defaults to "quasipoisson"}
}
\value{
List with meta-analysis results including AF and AN.
}
\description{
Performs meta-analysis across regions and calculates attributable
fraction and attributable number.
}
