% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrdl.R
\name{attrdl}
\alias{attrdl}
\title{FUNCTION FOR COMPUTING ATTRIBUTABLE MEASURES FROM DLNM}
\usage{
attrdl(
  x,
  basis,
  cases,
  model = NULL,
  coef = NULL,
  vcov = NULL,
  model.link = NULL,
  type = "af",
  dir = "back",
  tot = TRUE,
  cen,
  range = NULL,
  sim = FALSE,
  nsim = 5000
)
}
\arguments{
\item{x}{AN EXPOSURE VECTOR OR (ONLY FOR dir="back") A MATRIX OF LAGGED EXPOSURES}

\item{basis}{THE CROSS-BASIS COMPUTED FROM x}

\item{cases}{THE CASES VECTOR OR (ONLY FOR dir="forw") THE MATRIX OF FUTURE CASES}

\item{model}{THE FITTED MODEL}

\item{coef}{COEF FOR basis IF model IS NOT PROVIDED}

\item{vcov}{VCOV FOR basis IF model IS NOT PROVIDED}

\item{model.link}{LINK FUNCTION IF model IS NOT PROVIDED}

\item{type}{EITHER "an" OR "af" FOR ATTRIBUTABLE NUMBER OR FRACTION}

\item{dir}{EITHER "back" OR "forw" FOR BACKWARD OR FORWARD PERSPECTIVES}

\item{tot}{IF TRUE, THE TOTAL ATTRIBUTABLE RISK IS COMPUTED}

\item{cen}{THE REFERENCE VALUE USED AS COUNTERFACTUAL SCENARIO}

\item{range}{THE RANGE OF EXPOSURE. IF NULL, THE WHOLE RANGE IS USED}

\item{sim}{IF SIMULATION SAMPLES SHOULD BE RETURNED. ONLY FOR tot=TRUE}

\item{nsim}{NUMBER OF SIMULATION SAMPLES}
}
\value{
Attributable Numbers and Fractions
}
\description{
A function to calculate attributable numbers and fractions derived from
(c) Antonio Gasparrini 2015-2017.
}
\keyword{internal}
