% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarrhea.R
\name{attribution_calculation}
\alias{attribution_calculation}
\title{Attribution calculation for maximum temperature}
\usage{
attribution_calculation(
  data,
  param_term,
  model,
  level,
  param_threshold = 1,
  filter_year = NULL
)
}
\arguments{
\item{data}{Data list from combine_health_climate_data() function.}

\item{param_term}{A character vector or list containing parameter terms such
as "tmax" (maximum temperature) and "rainfall" (precipitation).
Default to "tmax"}

\item{model}{The fitted model from run_inla_models() function.}

\item{level}{Character. The spatial disaggregation level.
Can take one of the following values: "country", "region", or "district".}

\item{param_threshold}{Numeric. Threshold above which exposure is considered,
"attributable". Can take floats. Defaults to 1.}

\item{filter_year}{Integer. The year to filter to data to. Defaults to NULL.}
}
\value{
Results containing the attributable number and fraction at the chosen
dissagregation level.
}
\description{
The attribution calculation uses the attrdl function from
Gasparini and DNLM package
}
