% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{calculate_min_mortality_temp}
\alias{calculate_min_mortality_temp}
\title{Calculate minimum mortality values}
\usage{
calculate_min_mortality_temp(df_list, regions_df, blup = NULL)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{regions_df}{A dataframe with two columns. Column 1 is abbreviated
region names. Column 2 is user-specified region names.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}
}
\value{
\itemize{
\item \code{mintempregions_} A named numeric vector.
Minimum (optimum) mortality temperature per region.
}
}
\description{
Calculate the temperature at which there is minimum mortality
using the product of the basis matrix and blup
}
