% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{calculate_min_mortality_temp}
\alias{calculate_min_mortality_temp}
\title{Calculate minimum mortality values}
\usage{
calculate_min_mortality_temp(
  df_list,
  blup = NULL,
  independent_cols = NULL,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas
)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{independent_cols}{column name (or list of names) of extra independent
variable to include in regression (excluding temperature). Defaults to NULL.}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}
}
\value{
\itemize{
\item \code{mintempregions_} A named numeric vector.
Minimum (optimum) mortality temperature per region.
\item \code{an_thresholds} A dataframe with the optimal temperature range and
temperature thresholds for calculation of attributable deaths.
}
}
\description{
Calculate the temperature at which there is minimum mortality
using the product of the basis matrix and blup
}
\keyword{internal}
