% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{calculate_min_mortality_temp}
\alias{calculate_min_mortality_temp}
\title{Calculate minimum mortality values}
\usage{
calculate_min_mortality_temp(
  df_list,
  blup = NULL,
  independent_col,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas
)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}

\item{indepedent_col}{column names of independent
variables to include in regression (excluding temperature,
see config file for formula structure)}
}
\value{
\itemize{
\item \code{mintempregions_} A named numeric vector.
Minimum (optimum) mortality temperature per region.
}
}
\description{
Calculate the temperature at which there is minimum mortality
using the product of the basis matrix and blup
}
