% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{calculate_qaic}
\alias{calculate_qaic}
\title{QAIC calculation}
\usage{
calculate_qaic(data, print_results = FALSE)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data from which to fit models.}

\item{print_results}{Logical. Whether or not to print model summaries and
pearson dispersion statistics. Defaults to FALSE.}
}
\value{
Dataframe containing QAIC results for each lag.
}
\description{
QAIC calculation
}
\keyword{internal}
