% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{check_diseases_vif}
\alias{check_diseases_vif}
\title{Check multicollinearity using VIF and Condition Number}
\usage{
check_diseases_vif(data, inla_param, basis_matrices_choices, case_type)
}
\arguments{
\item{data}{A data frame from \code{combined_health_climate_data()} function,
containing the columns: \code{tmax},\code{rainfall}, \code{r_humidity}, \code{runoff}, \code{tmin},\code{ndvi},
and must be compatible with \code{set_cross_basis()} for generating DLNM matrices.}

\item{inla_param}{Character vector of parameter names representing all
climate variables to consider excluding the \code{basis_matrices_choices} parameter.}

\item{basis_matrices_choices}{Character vector specifying the main exposure variables
that should be included as DLNM basismatrices, and should be excluded from the \code{inla_param}.
It might be \code{tmax}for temperature exposure and \code{rainfall} if rainfall exposure.}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of 'diarrhea' or 'malaria'.}
}
\value{
A list with:
\describe{
\item{vif}{Named numeric vector of VIF values.}
\item{condition_number}{The condition number (Kappa) of the design matrix.}
\item{interpretation}{A qualitative interpretation of collinearity level.}
}
}
\description{
This function checks for multicollinearity among DLNM basis variables
(specifically tmax and rainfall) and selected confounders (r_humidity, runoff,
tmin, ndvi) using the Variance Inflation Factor (VIF) and condition number (Kappa).
}
\keyword{internal}
