% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatehealth-package.R
\docType{package}
\name{climatehealth-package}
\alias{climatehealth}
\alias{climatehealth-package}
\title{climatehealth: Package for Calculating Climate-Health Indicators}
\description{
R code allowing users to calculate climate-health indicators.
}
\author{
\strong{Maintainer}: Charlie Browning \email{charlie.browning@ons.gov.uk}

}
\keyword{internal}
