% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatehealth-package.R
\docType{package}
\name{climatehealth-package}
\alias{climatehealth}
\alias{climatehealth-package}
\title{climatehealth: Measuring the climate-health relationship}
\description{
\subsection{Overview}{

This package provides a suite of analysis functions for measuring the relationship
between various climate factors (indicators) and health outcomes.
}

\subsection{Included Indicators}{
\itemize{
\item Heat and Cold
\item Wildfires
\item Mental Health
\item Air Pollution
\item Waterborne Diseases
\item Vectorborne Diseases
}
}

\subsection{Author}{

Charlie Browning
}

\subsection{Maintainer}{

Charlie Browning \href{mailto:charlbrowning@ons.gov.uk}{charlbrowning@ons.gov.uk}
}

\subsection{License}{

MIT
}

\subsection{See Also}{

TBC
}
}
\author{
\strong{Maintainer}: Charlie Browning \email{charlie.browning@ons.gov.uk}

}
