% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatehealth-package.R
\docType{package}
\name{climatehealth-package}
\alias{climatehealth}
\alias{climatehealth-package}
\title{climatehealth: Measuring the climate-health relationship}
\description{
\subsection{Overview}{

This package provides a suite of analysis functions for measuring the relationship
between various climate factors (indicators) and health outcomes.
}

\subsection{Included Indicators}{
\itemize{
\item Heat and Cold
\item Wildfires
\item Mental Health
\item Air Pollution
\item Waterborne Diseases
\item Vectorborne Diseases
}
}

\subsection{License}{

MIT
}

\subsection{See Also}{

TBC
}
}
\author{
\strong{Maintainer}: Kenechi Omeke \email{kenechi.omeke@ons.gov.uk}

Authors:
\itemize{
  \item Charlie Browning \email{charlie.browning@ons.gov.uk}
  \item Etse Yawo Dzakpa
  \item Gladin Jose
  \item Matt Pearce
  \item Ellie Watkins
  \item Claire Hunt
  \item Beatrice Byukusenge
  \item Cassien Habyarimana
  \item Venuste Nyagahakwa
  \item Felix Scarbrough
  \item Treesa Shaji
  \item Bonnie Lewis
  \item Maquines Odhiambo Sewe
  \item Vijendra Ingole
}

Other contributors:
\itemize{
  \item Joe Panes [contributor]
  \item Wellcome [funder]
  \item Office for National Statistics (SOSCHI Project) [copyright holder]
}

}
