% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatehealth-package.R
\docType{package}
\name{climatehealth-package}
\alias{climatehealth}
\alias{climatehealth-package}
\title{climatehealth: Measuring the climate-health relationship}
\description{
\subsection{Overview}{

This package provides a suite of analysis functions for measuring the relationship
between various climate factors (indicators) and health outcomes.
}

\subsection{Included Indicators}{
\itemize{
\item Mortality attributable to high and low outdoor temperatures
\item Mortality attributable to wildfire-related PM2.5
\item Suicides attributable to extreme heat
\item Mortality attributable to short-term exposure to outdoor PM2.5 exposure
\item Diarrhea cases attributable to extreme temperatures and rainfall
\item Malaria cases attributable to extreme temperatures and  rainfall
}
}

\subsection{License}{

MIT
}

\subsection{The full range of topics include}{
\itemize{
\item Temperature-related health effects
\item Health effects of wildfires
\item Mental Health
\item Health effects of air pollution
\item Water-borne diseases
\item Vector-borne diseases
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://climate-health.officialstatistics.org}
}

}
\author{
\strong{Maintainer}: Kenechi Omeke \email{climate.health@ons.gov.uk}

Authors:
\itemize{
  \item Charlie Browning
  \item Etse Yawo Dzakpa
  \item Gladin Jose
  \item Matt Pearce
  \item Ellie Watkins
  \item Claire Hunt
  \item Beatrice Byukusenge
  \item Cassien Habyarimana
  \item Venuste Nyagahakwa
  \item Felix Scarbrough
  \item Treesa Shaji
  \item Bonnie Lewis
  \item Maquines Odhiambo Sewe
  \item Vijendra Ingole
}

Other contributors:
\itemize{
  \item Antony Brown [contributor]
  \item Euan Soutter [contributor]
  \item Gillian Flower [contributor]
  \item David Furley [contributor]
  \item Joe Panes [contributor]
  \item Charlotte Romaniuk [contributor]
  \item Milly Powell [contributor]
  \item Wellcome [funder]
  \item Office for National Statistics (SOSCHI Project) [copyright holder]
}

}
