% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{common_descriptive_stats}
\alias{common_descriptive_stats}
\title{The wrapper function to compute descriptive stats for the heat and cold indicator.}
\usage{
common_descriptive_stats(
  dataset_title,
  df_list,
  output_path,
  plot_corr_matrix = F,
  correlation_method = "pearson",
  plot_dist = F,
  dist_columns = c(),
  plot_ma = F,
  ma_days = 100,
  ma_sides = 2,
  ma_columns = c(),
  dependent_col,
  independent_cols,
  plot_na_counts = F,
  plot_scatter = F
)
}
\arguments{
\item{dataset_title}{String. The title for the dataset being used.}

\item{df_list}{list(dataframe) A list of input dataframes.}

\item{output_path}{string. The path to write outputs to.}

\item{correlation_method}{string. The correlation method used in ds.}

\item{dist_columns}{vector. The columns to plot distributions for.}

\item{ma_days}{int. The number of days to use for a moving average.}

\item{ma_sides}{int. The number of sides to use for a moving average (1 or 2).}

\item{ma_columns}{vector. Additional columns to plot moving average for. Dependent done by default.}

\item{dependent_col}{str. The column in the data containing the dependent var.}

\item{independent_cols}{str. The column in the data containing the independent var.}

\item{use_individual_dfs}{bool. Whether or not to use all dfs or concat them into one.}

\item{title}{string. title used for the plots in the ds.}
}
\description{
The wrapper function to compute descriptive stats for the heat and cold indicator.
}
