% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{common_descriptive_stats}
\alias{common_descriptive_stats}
\title{Deprecated alias for \code{run_descriptive_stats()}.}
\usage{
common_descriptive_stats(
  df_list,
  output_path,
  aggregation_column = NULL,
  population_col = NULL,
  plot_corr_matrix = FALSE,
  correlation_method = "pearson",
  plot_dist = FALSE,
  plot_ma = FALSE,
  ma_days = 100,
  ma_sides = 1,
  timeseries_col = NULL,
  dependent_col,
  independent_cols,
  units = NULL,
  plot_na_counts = FALSE,
  plot_scatter = FALSE,
  plot_box = FALSE,
  plot_seasonal = FALSE,
  plot_regional = FALSE,
  plot_total = FALSE,
  detect_outliers = FALSE,
  calculate_rate = FALSE
)
}
\arguments{
\item{df_list}{List of dataframes. A list of input dataframes.}

\item{output_path}{Character. The path to write outputs to.}

\item{aggregation_column}{Character. The column to use for aggregating the dataset into smaller subsets of regions.}

\item{population_col}{Character. The column containing the population.}

\item{plot_corr_matrix}{Logical. Whether or not to plot correlation matrix.}

\item{correlation_method}{Character. The correlation method. One of 'pearson', 'spearman', 'kendall'.}

\item{plot_dist}{Logical. Whether or not to plot distribution histograms.}

\item{plot_ma}{Logical. Whether to plot moving averages over a timeseries.}

\item{ma_days}{Integer. The number of days to use for a moving average.}

\item{ma_sides}{Integer. The number of sides to use for a moving average (1 or 2).}

\item{timeseries_col}{Character. The column used as the timeseries for moving averages.}

\item{dependent_col}{Character. The column in the data containing the dependent variable.}

\item{independent_cols}{Character vector. The columns in the data containing the independent variables.}

\item{units}{Named character vector. A named character vector of units for each variable.}

\item{plot_na_counts}{Logical. Whether to plot NA counts.}

\item{plot_scatter}{Logical. Whether to plot scatter plots.}

\item{plot_box}{Logical. Whether to plot box plots.}

\item{plot_seasonal}{Logical. Whether to plot seasonal plots.}

\item{plot_regional}{Logical. Whether to plot regional plots.}

\item{plot_total}{Logical. Whether to plot total health outcomes per year.}

\item{detect_outliers}{Logical. Whether to output a table containing outlier information.}

\item{calculate_rate}{Logical. Whether to calculate the rate of health outcomes per 100k people.}
}
\value{
Character vector. Backward-compatible output path format.
}
\description{
Generic wrapper function to compute descriptive statistics and EDA outputs.
}
\keyword{internal}
