% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{common_descriptive_stats_api}
\alias{common_descriptive_stats_api}
\title{The function used to create desciprtive stats via an API endpoint}
\usage{
common_descriptive_stats_api(
  data,
  aggregation_column = NULL,
  dataset_title,
  dependent_col,
  independent_cols,
  plot_correlation = T,
  plot_dist_hists = T,
  plot_ma = T,
  plot_na_counts = T,
  plot_scatter = T,
  correlation_method = NULL,
  dist_columns = NULL,
  ma_days = NULL,
  ma_sides = 2,
  ma_columns = NULL,
  timeseries_col = NULL,
  output_path
)
}
\arguments{
\item{data}{The dataset used for descriptive stats (as a vector)}

\item{aggregation_column}{The column to use for aggregating the dataset into smaller subsets}

\item{dataset_title}{The datasets title used for outputs}

\item{dependent_col}{The dependent column}

\item{independent_cols}{A vector of independent columns}

\item{plot_correlation}{Whether to plot a correlation matrix}

\item{plot_dist_hists}{Whether to plot histograms showing column distributions}

\item{plot_ma}{Whether to plot moving averages over a timeseries}

\item{plot_na_counts}{Whether to plot counts of NAs in each column}

\item{plot_scatter}{Whether to plot the dependent column against the independent columns}

\item{correlation_method}{The correlation method. One of 'pearson', 'spearman', 'kendall'}

\item{dist_columns}{The columns to plot distributions for}

\item{ma_days}{The number of days to use in moving average calculations}

\item{ma_sides}{The number of sides to use in moving average calculations (1 or2)}

\item{ma_columns}{A vector of columns to plot moving average for}

\item{timeseries_col}{The column used as the timerseries for moving averages}

\item{output_path}{The path to save outputs to}
}
\value{
The full directory path that the descriptive stats are saved to.
}
\description{
The function used to create desciprtive stats via an API endpoint
}
