% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{common_descriptive_stats_core}
\alias{common_descriptive_stats_core}
\title{Deprecated alias for \code{descriptive_stats_core()}.}
\usage{
common_descriptive_stats_core(
  df,
  output_path,
  title,
  aggregation_column = NULL,
  population_col = NULL,
  plot_corr_matrix = FALSE,
  correlation_method = "pearson",
  plot_dist = FALSE,
  dependent_col,
  independent_cols = c(),
  units = NULL,
  plot_na_counts = FALSE,
  plot_scatter = FALSE,
  plot_box = FALSE,
  plot_seasonal = FALSE,
  plot_regional = FALSE,
  plot_total = FALSE,
  timeseries_col = "date",
  detect_outliers = FALSE,
  calculate_rate = FALSE
)
}
\description{
Deprecated. Use \code{descriptive_stats_core()} instead.
}
\keyword{internal}
