% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{common_descriptive_stats_core}
\alias{common_descriptive_stats_core}
\title{The core functionality for producing descriptive stats}
\usage{
common_descriptive_stats_core(
  dataset_title,
  df,
  output_path,
  title,
  plot_corr_matrix = F,
  correlation_method = "pearson",
  plot_dist = F,
  dist_columns = c("temp"),
  dependent_col = "dependent",
  independent_cols = c("temp"),
  plot_na_counts = F,
  plot_scatter = F
)
}
\arguments{
\item{dataset_title}{String. The title for the dataset being used.}

\item{df}{DataFrame. Input data for descriptive stats/}

\item{output_path}{string. The path to write outputs to.}

\item{title}{string. title used for the plots in the ds.}

\item{correlation_method}{string. The correlation method used in ds.}

\item{dist_columns}{vector. The columns to plot distributions for.}

\item{dependent_col}{str. The column in the data containing the dependent var.}

\item{independent_cols}{str. The column in the data containing the independent var.}
}
\description{
The core functionality for producing descriptive stats
}
