% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{contour_plot}
\alias{contour_plot}
\title{Create a contour plot at country, region or district level.}
\usage{
contour_plot(
  data,
  param_term,
  model,
  level,
  filter_year = NULL,
  case_type,
  save_fig = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{data}{Data list from combine_health_climate_data() function.}

\item{param_term}{A character vector or list containing parameter terms such
as "tmax" (maximum temperature) and "rainfall" (precipitation).
Default to "tmax"}

\item{model}{The fitted model from run_inla_models() function.}

\item{level}{A character vector specifying the spatial disaggregation level.
Can take one of the following values: "country", "region", or "district".}

\item{filter_year}{Integer. The year to filter to data to. Defaults to NULL.}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of 'diarrhea' or 'malaria'}

\item{save_fig}{Boolean. Whether to save the outputted plot. Defaults to
FALSE.}

\item{output_dir}{The path to save the visualisation to. Defaults to NULL}
}
\value{
contour plot at country, Region and District level
}
\description{
: Generates a contour plot showing the lag exposure effect of
maximum temperature (tmax) and cumulative rainfall on Malaria cases.
}
\keyword{internal}
