% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{create_grid}
\alias{create_grid}
\title{Generate a grid size for a certain number of plots.}
\usage{
create_grid(plot_count)
}
\arguments{
\item{plot_count}{The number of plots required for the grid.}
}
\value{
A numeric vector: c(x, y), where x and y define the grid dimensions.
}
\description{
This function calculates the minimum grid size required to plot X amount of
plots on a a figure. For example, 6 plots would require a 3x2, where as 7
would require a 3x3, and so on.
}
