% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{create_inla_indices}
\alias{create_inla_indices}
\title{Create indices for INLA models}
\usage{
create_inla_indices(data, case_type)
}
\arguments{
\item{data}{is the dataset containing district_code, region_code, and year
columns from the combine_health_climate_data() function.}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of 'diarrhea' or 'malaria'}
}
\value{
The modified data with the created indices.
}
\description{
: For the INLA model, there is a need to set-up regions index,
district index, and year index. This function create these indices using the
dataset, ndistrict and nregion defined above.
}
\keyword{internal}
