% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{create_temperature_splines}
\alias{create_temperature_splines}
\title{Generate splines for temperature variable}
\usage{
create_temperature_splines(data, nlag = 0, degrees_freedom = 6)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data}

\item{nlag}{Integer. The number of days of lag in the temperature
variable from which to generate splines. Default is 0 (unlagged temperature
variable). Defaults to 0.}

\item{degrees_freedom}{Integer. Degrees of freedom for the spline(s).
Defaults to 6.}
}
\value{
Dataframe with additional column for temperature spline.
}
\description{
Generates temperature splines for each region
}
\keyword{internal}
