% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{create_temperature_splines}
\alias{create_temperature_splines}
\title{Generate splines for temperature variable}
\usage{
create_temperature_splines(data, temperature_lag = 0, degrees_freedom = 6)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data}

\item{temperature_lag}{Integer. The number of days of lag in the temperature
variable from which to generate splines. Default is 0 (unlagged temperature
variable).}

\item{df}{Integer. Degrees of freedom for the spline(s).}
}
\value{
Dataframe with additional columns for temperature spline
}
\description{
Generates temperature splines for each region
}
