% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{define_and_validate_optimal_temps}
\alias{define_and_validate_optimal_temps}
\title{Define and validate the optimal temperature range from the model predictions.}
\usage{
define_and_validate_optimal_temps(optimal_temp_range, prediction, index)
}
\arguments{
\item{optimal_temp_range}{An matrix used to store the optimal temperature
ranges.}

\item{prediction}{The models prediction}
}
\value{
The optimal temperature range.
}
\description{
Define and validate the optimal temperature range from the model predictions.
}
