% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{define_model}
\alias{define_model}
\title{Define regression model}
\usage{
define_model(
  dataset,
  independent_col,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas
)
}
\arguments{
\item{dataset}{dataframe with tmean column to be modelled}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}

\item{indepedent_col}{column names of independent
variables to include in regression (excluding temperature,
see config file for formula structure)}
}
\value{
\itemize{
\item \code{model} A quasi-poission generalised linear model object.
See: https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/glm
\item \code{cb} Basis matrices for the two dimensions of predictor and lags.
}
}
\description{
Define regression model
}
