% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect Outliers Using the IQR Method}
\usage{
detect_outliers(df, columns = NULL)
}
\arguments{
\item{df}{A data frame containing the data to check for outliers.}

\item{columns}{A character vector of column names to check for outliers.}
}
\value{
Dataframe. Column summaries
}
\description{
Detect Outliers Using the IQR Method
}
