% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnm_shared.R
\name{dlnm_predict_nat}
\alias{dlnm_predict_nat}
\title{Run national predictions from meta analysis}
\usage{
dlnm_predict_nat(
  df_list,
  var_fun = "bs",
  var_per = c(25, 50, 75),
  var_degree = 2,
  minpercreg,
  mmpredall,
  pred_list,
  country = "National"
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(25,50,75).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{minpercreg}{Vector. Percentile of maximum suicide temperature for each region.}

\item{mmpredall}{List of national coefficients and covariance matrices for the crosspred.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{country}{Character. Name of country for national level estimates.}
}
\value{
A list containing predictions by region.
}
\description{
Use the meta analysis to create national level predictions
}
\keyword{internal}
