% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{do_analysis}
\alias{do_analysis}
\title{Do full dlnm analysis}
\usage{
do_analysis(input_csv_path, output_csv_path)
}
\arguments{
\item{output_csv_path}{}
}
\value{
a plot of temperature versus relative risk for
each region and a CSV of the data used to make the plot
}
\description{
Do full dlnm analysis
}
