% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{do_analysis}
\alias{do_analysis}
\title{Do full DLNM analysis}
\usage{
do_analysis(
  input_csv_path_,
  output_folder_path_,
  save_fig_ = TRUE,
  save_csv_ = TRUE,
  meta_analysis,
  dependent_col_,
  independent_col_,
  time_col_,
  region_col_,
  temp_col_,
  varfun_,
  varper_,
  vardegree_,
  lag_,
  lagnk_,
  dfseas_
)
}
\arguments{
\item{meta_analysis}{Boolean. Whether to include meta-analysis.
TRUE or FALSE.}

\item{input_csv_path}{Path to a CSV contain
daily time series of death and temperature per region.}

\item{output_folder_path}{Path to folder for storing outputs.}
}
\value{
A PDF containing a line plot of temperature versus
relative risk per region,
and histogram of temperatures per region.
A CSV of relative risk per temperature per region.
}
\description{
Runs a sequence of functions to carry out
heat-related mortality analysis.
}
\details{
Modified from Gasparrini A et al. (2015)
The Lancet. 2015;386(9991):369-375.
}
\seealso{
\link{dlnm} package
}
