% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{do_analysis}
\alias{do_analysis}
\title{Do full DLNM analysis (modified from Gasparrini et al. 2015)}
\usage{
do_analysis(input_csv_path, output_folder_path)
}
\arguments{
\item{input_csv_path}{Path to a CSV contain daily time series of death and temperature per region.}

\item{output_folder_path}{Path to folder for storing outputs.}
}
\value{
A PDF containing a line plot of temperature versus relative risk per region,
and histogram of temperatures per region. A CSV of  relative risk per temperature per region.
}
\description{
Do full DLNM analysis (modified from Gasparrini et al. 2015)
}
