% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{do_analysis}
\alias{do_analysis}
\title{Do full DLNM analysis}
\usage{
do_analysis(
  input_csv_path_,
  output_folder_path_,
  save_fig_,
  save_csv_,
  meta_analysis,
  by_region,
  time_range_,
  dependent_col_,
  independent_col_,
  time_col_,
  region_col_,
  temp_col_,
  varfun_,
  varper_,
  vardegree_,
  lag_,
  lagnk_,
  dfseas_
)
}
\arguments{
\item{input_csv_path_}{Path to a CSV contain
daily time series of death and temperature per region.}

\item{output_folder_path_}{Path to folder for storing outputs.}

\item{save_fig_}{Boolean (TRUE or FALSE). Whether to save output figure.}

\item{save_csv_}{Boolean (TRUE or FALSE). Whether to save output CSVs.}

\item{meta_analysis}{Boolean (TRUE or FALSE). Whether to include
meta-analysis. Must be TRUE if by_region argument is FALSE.}

\item{by_region}{Boolean (TRUE or FALSE). Whether to disaggregate by region.
Must be TRUE if meta-analysis is FALSE.}

\item{time_range_}{Time range over which to run the analysis.}

\item{dependent_col_}{the column name of the
dependent variable of interest e.g. deaths}

\item{time_col_}{The column name of column containing dates (e.g date, year).}

\item{region_col_}{The column name of the column containing regions.}

\item{temp_col_}{the column name of the column containing the exposure.}

\item{varfun_}{Exposure function
(see dlnm::crossbasis)}

\item{varper_}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree_}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag_}{Lag length in time
(see dlnm::logknots)}

\item{lagnk_}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas_}{Degrees of freedom for seasonality}

\item{indepedent_col_}{column names of independent
variables to include in regression (excluding temperature,
see config file for formula structure)}
}
\value{
\itemize{

\item A PDF containing a line plot of temperature versus
relative risk per region,
and histogram of temperatures per region.
\item A CSV of relative risk per temperature per region.
\item \code{output_df} A dataframe with relative risk estimates and confidence
intervals across the temperature range for each region.
\item \code{tmean_df} A dataframe with daily mean exposure values for each
region.
\item \code{anregions_bind} A matrix of numbers of deaths attributable to
temperature, heat, cold, extreme heat and extreme cold (with confidence
intervals), disaggregated by region.
\item \code{attrdl_yr_all} A dataframe with attributable deaths by year for each
region.
\item \code{attr_fractions_yr} A dataframe with attributable fractions by year
and region.
}
}
\description{
Runs a sequence of functions to carry out
heat-related mortality analysis.
}
\details{
Modified from Gasparrini A et al. (2015)
The Lancet. 2015;386(9991):369-375.
}
\seealso{
\link{dlnm} package
}
