% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{fit_air_pollution_gam}
\alias{fit_air_pollution_gam}
\title{Fit GAM model}
\usage{
fit_air_pollution_gam(data, var_name = "pm25", family = "quasipoisson")
}
\arguments{
\item{data}{Data to fit the model.}

\item{var_name}{Character. The name of the column containing PM2.5 values.
Defaults to "pm25"}

\item{family}{Character string indicating the distribution family used in the GAM.
Must be one of the four options "quasipoisson", "nb", "ziP", and "poisson".
"quasipoisson" (Appropriate for count data with overdispersion),
"nb" (negativebinomial which is suitable for highly dispersed count data),
"ziP" (zero-inflated which is suitable for count data with excess zeros (sparse data)),
"poisson" (Suitable for count data without overdispersion (mean = variance)).
Defaults to "quasipoisson"}
}
\value{
GAM model object or NULL if failed.
}
\description{
Fits the GAM model for a single region with a natural spline
for climate variables.
}
\keyword{internal}
