% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{fwald}
\alias{fwald}
\title{Calculate p-values for Wald test}
\usage{
fwald(mm, var)
}
\arguments{
\item{mm}{A model object. A multivariate meta-analysis model.}

\item{var}{A character. The name of the variable in the meta-model to
calculate p-values for.}
}
\value{
A number. The p-value of the explanatory variable.
}
\description{
Calculate p-values for an explanatory variable.
}
\keyword{internal}
