% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{fwald}
\alias{fwald}
\title{Calculate p-values for Wald test}
\usage{
fwald(mm, var)
}
\arguments{
\item{var}{A character. The name of the variable in the model to calculate
p-values for.}

\item{model}{A model object.}
}
\value{
A number. The p-value of the explanatory variable.
}
\description{
A function to calculate p-values for an explanatory variable.
}
\keyword{internal}
