% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soschi_theme.R
\name{get_accessible_palette}
\alias{get_accessible_palette}
\title{Get project plot palette}
\usage{
get_accessible_palette()
}
\value{
A named list of hex colour values.
}
\description{
Returns the standard project palette for accessible plots.
}
\keyword{internal}
