% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soschi_theme.R
\name{get_month_labels}
\alias{get_month_labels}
\title{Standard month labels}
\usage{
get_month_labels()
}
\value{
Character vector of month abbreviations.
}
\description{
Returns abbreviated month labels for use on x-axes.
}
\keyword{internal}
