% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soschi_theme.R
\name{get_plot_grid}
\alias{get_plot_grid}
\title{Get plot grid dimensions}
\usage{
get_plot_grid(n_plots, max_cols = 2)
}
\arguments{
\item{n_plots}{Integer. Number of plots.}

\item{max_cols}{Integer. Maximum number of columns.}
}
\value{
A list with \code{n_col} and \code{n_row}.
}
\description{
Calculates the number of columns and rows for a multi-panel plot layout.
Defaults to a maximum of 2 columns to avoid squished panels.
}
\keyword{internal}
