% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{get_predictions}
\alias{get_predictions}
\title{A function to predict relative risk at country, region, and district level}
\usage{
get_predictions(data, param_term, model, level, case_type)
}
\arguments{
\item{data}{Data list from combine_health_climate_data() function.}

\item{param_term}{A character vector or list containing parameter terms such
as "tmax" (maximum temperature) and "rainfall" (precipitation).}

\item{model}{The fitted model from run_inla_models() function.}

\item{level}{Character. The spatial disaggregation level.
Can take one of the following values: "country", "region", or "district".}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of 'diarrhea' or 'malaria'}
}
\value{
A dataframe containing cumulative relative risk at the chosen level.
}
\description{
Produces cumulative relative risk at country, region and
district level from analysis.
}
