% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{get_region_metadata}
\alias{get_region_metadata}
\title{Get region metadata for analysis}
\usage{
get_region_metadata(regions, df_list_unordered, region_names = NULL)
}
\arguments{
\item{regions}{A list of strings representing region names.
\strong{NOTE} Must be same order as input data.}

\item{df_list_unordered}{list of dataframes for each region}
}
\value{
\itemize{
\item regions. A dataframe with two columns. Column 1 is abbreviated
region names. Column 2 is full region names.
\item df_list. An alphabetically-ordered list of dataframes for each region.
}
\description{
Get region metadata for analysis
}
