% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{get_wildfire_lag_columns}
\alias{get_wildfire_lag_columns}
\title{Get wildfire lag columns}
\usage{
get_wildfire_lag_columns(data)
}
\arguments{
\item{data}{Dataframe containing a daily time series of climate and health
data with wildfire PM2.5 data}
}
\value{
List containing wildfire lag column names and lag numbers
}
\description{
Finds wildfire PM2.5 lag columns in a dataframe
}
\keyword{internal}
