% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_add_national_data}
\alias{hc_add_national_data}
\title{Process data for national analysis}
\usage{
hc_add_national_data(
  df_list,
  pop_list,
  var_fun = "bs",
  var_per = c(10, 75, 90),
  var_degree = 2,
  lagn = 21,
  lagnk = 3,
  country = "National",
  cb_list,
  mm,
  minpercgeog_
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{pop_list}{List of population totals by year and geography.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(10, 75, 90).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{lagn}{Integer. Number of days in the lag period. Defaults to 21.
(see dlnm::crossbasis).}

\item{lagnk}{Integer. Number of knots in lag function. Defaults to 3.
(see dlnm::logknots).}

\item{country}{Character. Name of country for national level estimates.
Defaults to 'National'.}

\item{cb_list}{A list of cross-basis matrices by geography.}

\item{mm}{A model object. A multivariate meta-analysis model.}

\item{minpercgeog_}{Vector. Percentile of minimum mortality temperature for
each geography.}
}
\value{
\itemize{
\item \code{df_list} List. A list of data frames for each geography and
national level.
\item \code{cb_list} List. A list of cross-basis matrices by geography and
national level.
\item \code{minpercgeog_} Vector. Percentile of minimum mortality temperature
for each geography and national level.
\item \code{mmpredall} List. A list of national coefficients and covariance
matrices.
}
}
\description{
Aggregate to national data and run crossbasis
}
\keyword{internal}
