% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_adf}
\alias{hc_adf}
\title{Run ADF test and produce PACF plots for each model combo}
\usage{
hc_adf(df_list)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular geography.}
}
\value{
adf_result for each geography.
}
\description{
Run augmented Dickey-Fuller test for stationarity of dependent variable
and produce partial autocorrelation function plot of residuals for each model combo
}
\keyword{internal}
