% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_attr}
\alias{hc_attr}
\title{Estimate attributable numbers}
\usage{
hc_attr(
  df_list,
  cb_list,
  pred_list,
  minpercgeog_,
  attr_thr_high = 97.5,
  attr_thr_low = 2.5,
  seed = NULL
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{cb_list}{A list of cross-basis matrices by geography.}

\item{pred_list}{A list containing predictions from the model by geography.}

\item{minpercgeog_}{Vector. Percentile of minimum mortality temperature for
each geography.}

\item{attr_thr_high}{Integer. Percentile at which to define the upper
temperature threshold for calculating attributable risk. Defaults to 97.5.}

\item{attr_thr_low}{Integer. Percentile at which to define the lower
temperature threshold for calculating attributable risk. Defaults to 2.5.}

\item{seed}{Optional integer random seed used when sampling residuals for
model validation plots. Defaults to NULL.}
}
\value{
'attr_list'. A list containing attributable numbers per geography.
}
\description{
Estimate attributable numbers and confidence intervals for
each geography using Monte Carlo simulations.
}
\keyword{internal}
