% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_attr}
\alias{hc_attr}
\title{Estimate attributable numbers}
\usage{
hc_attr(
  df_list,
  cb_list,
  pred_list,
  minpercgeog_,
  attr_thr_high = 97.5,
  attr_thr_low = 2.5
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{A list of cross-basis matrices by region.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{minpercreg}{Vector. Percentile of maximum suicide temperature for each region.}

\item{attr_thr}{Integer. Percentile at which to define the temperature threshold for
calculating attributable risk.}
}
\value{
A list containing attributable numbers per region
}
\description{
Estimate attributable numbers for each region and confidence
intervals using Monte Carlo simulations.
}
\keyword{internal}
