% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_plot_ar_cold_monthly}
\alias{hc_plot_ar_cold_monthly}
\title{Plot attributable rates by calendar month}
\usage{
hc_plot_ar_cold_monthly(
  attr_mth_list,
  df_list,
  country = "National",
  attr_thr_low = 2.5,
  save_fig = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{attr_mth_list}{A list of data frames containing total attributable
fractions, numbers and rates by calendar month and area.}

\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{country}{Character. Name of country for national level estimates.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}

\item{attr_thr}{Integer. Percentile at which to define the temperature threshold for
calculating attributable risk.}
}
\value{
Plots of attributable rates by calendar month per area
}
\description{
Plot attributable rates grouped over the whole time series by
calendar month to explore seasonality.
}
\keyword{internal}
