% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_plot_power}
\alias{hc_plot_power}
\title{Plot power calculations for temperature mortality analysis.}
\usage{
hc_plot_power(
  power_list_high,
  power_list_low,
  save_fig = FALSE,
  output_folder_path = NULL,
  country = "National"
)
}
\arguments{
\item{power_list_high}{List. High power values.}

\item{power_list_low}{List. Low power values.}

\item{save_fig}{Logical. Whether or not to save the figure. Defailts to FALSE.}

\item{output_folder_path}{Character. The output location for the figure. Defaults
to NULL.}

\item{country}{Character. The name of the country or group of regions of analysis.
Defaults to "National".}
}
\value{
Dataframe containing cumulative relative risk and confidence
intervals from analysis.
}
\description{
Produces a plot displaying power calculation values for
temperature mortatlity analysis.
}
\keyword{internal}
