% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_plot_rr}
\alias{hc_plot_rr}
\title{Plot results of relative risk analysis - Mental Health}
\usage{
hc_plot_rr(
  df_list,
  pred_list,
  attr_thr_high = 97.5,
  attr_thr_low = 2.5,
  minpercgeog_,
  country = "National",
  save_fig = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{attr_thr_high}{Integer. Percentile at which to define the upper temperature threshold for
calculating attributable risk. Defaults to 97.5.}

\item{attr_thr_low}{Integer. Percentile at which to define the lower temperature threshold for
calculating attributable risk. Defaults to 2.5.}

\item{minpercgeog_}{Vector. Percentile of minimum suicide temperature for each area.}

\item{country}{Character. Name of country for national level estimates.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots should be saved.
Defaults to NULL.}
}
\value{
Plots of cumulative lag exposure-response function with histogram of
temperature distribution for each region
}
\description{
Plots cumulative lag exposure-response function with histogram of
temperature distribution for each region
}
\keyword{internal}
