% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_predict_subnat}
\alias{hc_predict_subnat}
\title{Run predictions from model}
\usage{
hc_predict_subnat(
  df_list,
  var_fun = "bs",
  var_per = c(10, 75, 90),
  var_degree = 2,
  minpercgeog_,
  blup,
  coef_,
  vcov_,
  meta_analysis = FALSE
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(25,50,75).}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm::crossbasis). Defaults to 2 (quadratic).}

\item{blup}{A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each region.}

\item{coef_}{A matrix of coefficients for the reduced model.}

\item{vcov_}{A list. Covariance matrices for each region for the reduced model.}

\item{meta_analysis}{Boolean. Whether to perform a meta-analysis.}

\item{minpercreg}{Vector. Percentile of maximum suicide temperature for each region.}
}
\value{
A list containing predictions by region
}
\description{
Use model to run predictions. Predictions can be produced for a single input geography,
or multiple disaggregated geographies.
}
\keyword{internal}
