% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_mortality.R
\name{hc_quasipoisson_dlnm}
\alias{hc_quasipoisson_dlnm}
\title{Define and run quasi-Poisson regression with DLNM}
\usage{
hc_quasipoisson_dlnm(
  df_list,
  control_cols = NULL,
  cb_list,
  dfseas = 8,
  df_control = 3
)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a
health outcome and climate variables which may be disaggregated by a
particular geography.}

\item{control_cols}{List. Confounders to include in the final model
adjustment. Defaults to NULL.}

\item{cb_list}{List of cross-basis matrices from hc_create_crossbasis
function.}

\item{dfseas}{Integer. Degrees of freedom for seasonality. Defaults to 8.}

\item{df_control}{Integer. Degrees of freedom for natural splines applied to
control variables. Defaults to 3.}
}
\value{
'model_list'. List containing models by geography.
}
\description{
Fits a quasi-Poisson time series regression with a
distributed lag non-linear model.
}
\keyword{internal}
