% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_read_data}
\alias{hc_read_data}
\title{Read heat and cold indicator data}
\usage{
hc_read_data(
  input_csv_path,
  dependent_col,
  date_col,
  region_col,
  temperature_col,
  population_col
)
}
\arguments{
\item{input_csv_path}{Path to a CSV containing a
daily time series of death counts and temperature per geography.}

\item{dependent_col}{the column name of the health outcome
dependent variable of interest e.g,. deaths.}

\item{date_col}{Date column.}

\item{population_col}{The population estimate column e.g., pop.}

\item{geog}{The geography column over which the data
are spatially aggregated e.g., geognames.}

\item{temp_col}{The temperature variable column e.g., tmean.}
}
\value{
An alphabetically-ordered list of dataframes for each
geography comprising of dates, deaths, and temperatures.
}
\description{
Reads in data and geography names for analysis from a CSV file.
}
\keyword{internal}
