% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_rr_results}
\alias{hc_rr_results}
\title{Produce cumulative relative risk results of analysis}
\usage{
hc_rr_results(pred_list, df_list, attr_thr_high = 97.5, attr_thr_low = 2.5)
}
\arguments{
\item{pred_list}{A list containing predictions from the model by region.}
}
\value{
Dataframe containing cumulative relative risk and confidence
intervals from analysis.
}
\description{
Produces cumulative relative risk and confidence intervals
from analysis.
}
\keyword{internal}
