% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{hc_save_results}
\alias{hc_save_results}
\title{Save results of analysis - Mental Health}
\usage{
hc_save_results(
  rr_results,
  res_attr_tot,
  attr_yr_list,
  attr_mth_list,
  power_list_high,
  power_list_low,
  output_folder_path = NULL
)
}
\arguments{
\item{rr_results}{Dataframe containing cumulative relative risk and confidence
intervals from analysis.}

\item{res_attr_tot}{Matrix containing total attributable fractions, numbers and rates for each
area over the whole time series.}

\item{attr_yr_list}{A list of matrices containing yearly estimates of attributable
fractions, numbers and rates by area}

\item{attr_mth_list}{A list of data frames containing total attributable
fractions, numbers and rates by calendar month and area.}

\item{output_folder_path}{Path to folder where results should be saved.
Defaults to NULL.}
}
\description{
Saves a CSV file of cumulative relative risk and
confidence intervals.
}
\keyword{internal}
