% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{load_air_pollution_data}
\alias{load_air_pollution_data}
\title{R-code for I1: All-cause mortality attributable to Outdoor PM2.5.
Read in climate, environmental and health data and rename columns}
\usage{
load_air_pollution_data(
  data_path,
  date_col = "date",
  region_col = "province",
  pm25_col = "pm25",
  deaths_col = "deaths",
  humidity_col = "humidity",
  precipitation_col = "precipitation",
  tmax_col = "tmax",
  population_col = NULL,
  age_col = NULL,
  sex_col = NULL,
  urbanisation_col = NULL
)
}
\arguments{
\item{data_path}{Path to a CSV file containing a daily time series of data.}

\item{date_col}{Character. Name of date column in the dataframe with format
YYYY-MM-DD. Defaults to "date".}

\item{region_col}{Character. Name of region column in the dataframe. Defaults
to "province".}

\item{pm25_col}{Character. Name of PM2.5 column in the dataframe. Defaults to
"pm25".}

\item{deaths_col}{Character. Name of all-cause mortality column in the dataframe.
Defaults to "deaths"}

\item{humidity_col}{Character. Name of humidity column in the dataframe. Defaults
to "humidity".}

\item{precipitation_col}{Character. Name of precipitation column in the dataframe.
Defaults to "precipitation".}

\item{tmax_col}{Character. Name of temperature column in the dataframe.
Defaults to "tmax".}

\item{population_col}{Character. Name of population column in the dataframe.
Defaults to NULL.}

\item{age_col}{Character. Name of age column in the dataframe. Defaults to NULL.}

\item{sex_col}{Character. Name of sex column in the dataframe. Defaults to NULL.}

\item{urbanisation_col}{Character. Name of urbanisation column in the dataframe.
Defaults to NULL.}
}
\value{
Dataframe with formatted and renamed with standardized column names.
}
\description{
Reads in a CSV file for a daily time series of climate, environmental
and health data and renames them to standardised names. This function
creates year, month, day, and day of week columns derived from the date.
}
