% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{load_and_process_data}
\alias{load_and_process_data}
\title{Read in and format health data - diseases cases type}
\usage{
load_and_process_data(
  health_data_path,
  region_col,
  district_col,
  date_col = NULL,
  year_col = NULL,
  month_col = NULL,
  case_col,
  case_type,
  tot_pop_col
)
}
\arguments{
\item{health_data_path}{Path to a csv file containing a monthly time series of data
for health outcome case type, which may be disaggregated by sex (under five case or
above five case), and by Region and District.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names.}

\item{district_col}{Character. Name of the column in the dataframe that
contains the district names.}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date. Defaults to NULL.}

\item{year_col}{Character. Name of the column in the dataframe that contains
the year.}

\item{month_col}{Character. Name of the column in the dataframe that contains
the month.}

\item{case_col}{Character. Name of the column in the dataframe
that contains the disease cases to be considered.}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of 'diarrhea' or 'malaria'.}

\item{tot_pop_col}{Character. Name of the column in the dataframe that contains
the total population.}
}
\value{
A dataframe with formatted and renamed columns.
}
\description{
Read in a csv file containing a monthly time series of health
outcomes and population data. Renames columns and creates time variables for
spatiotemporal analysis.
}
\keyword{internal}
