% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{load_data}
\alias{load_data}
\title{Load data for analysis}
\usage{
load_data(
  input_csv_path,
  dependent_col,
  time_col,
  region_col,
  temp_col,
  population_col,
  output_year,
  RR_distribution_length
)
}
\arguments{
\item{input_csv_path}{Path to a CSV containing a
daily time series of death and temperature per region.}

\item{dependent_col}{the column name of the
dependent variable of interest e.g. deaths}

\item{time_col}{Time column e.g. date}

\item{region_col}{The region column over which the data
are spatially aggregated e.g. regnames}

\item{temp_col}{The temperature column e.g. tmean}

\item{population_col}{The population column e.g. pop}

\item{RR_distribution_length}{Number of years for the calculation of RR distribution. Set both as 'NONE' to use full range in data.}

\item{output_year_}{Year(s) to calculate output for.}
}
\value{
\code{df_list} An alphabetically-ordered list of dataframes for each
region comprising dates, deaths, and temperatures.
}
\description{
Loads data and names of regions for analysis from a CSV file.
}
