% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{load_data}
\alias{load_data}
\title{Load data for analysis}
\usage{
load_data(input_path)
}
\arguments{
\item{input_path}{a path to CSV}
}
\value{
\itemize{
\item df_list_unordered - a list of dataframes for each region
\item regions - a list of strings of the names of each region
}
\description{
Load data for analysis
}
