% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{load_data}
\alias{load_data}
\title{Load data for analysis}
\usage{
load_data(input_csv_path, dependent_col, time_col, region_col, temp_col)
}
\arguments{
\item{input_path}{Path to a CSV containing a
daily time series of death and temperature per region.}
}
\value{
\itemize{
\item \code{df_list_unordered} A list of dataframes for each region
comprising dates, deaths, and temperatures.
\item \code{regions} A character vector with the names of each region.
}
}
\description{
Loads data and names of regions for analysis from a CSV file.
}
