% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_attr}
\alias{mh_attr}
\title{Estimate attributable numbers}
\usage{
mh_attr(df_list, cb_list, pred_list, minpercreg, attr_thr = 97.5, seed = NULL)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{A list of cross-basis matrices by region.}

\item{pred_list}{A list containing predictions from the model by region.}

\item{minpercreg}{Vector. Percentile of maximum suicide temperature for each region.}

\item{attr_thr}{Integer. Percentile at which to define the temperature threshold for
calculating attributable risk. Defaults to 97.5.}

\item{seed}{Optional integer random seed used when sampling residuals for
model validation plots. Defaults to NULL.}
}
\value{
A list containing attributable numbers per region
}
\description{
Estimate attributable numbers for each region and confidence
intervals using Monte Carlo simulations.
}
\keyword{internal}
