% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_create_crossbasis}
\alias{mh_create_crossbasis}
\title{Create cross-basis matrix}
\usage{
mh_create_crossbasis(
  data,
  var_fun = "ns",
  var_dof = 4,
  lag_fun = "ns",
  lag_dof = 4,
  lag_days = 3
)
}
\arguments{
\item{data}{Dataframe containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'ns'.}

\item{var_dof}{Integer. Degrees of freedom in exposure function for argvar
(see dlnm:crossbasis). Defaults to 4.}

\item{lag_fun}{Character. Exposure function for arglag
(see dlnm::crossbasis). Defaults to 'ns'.}

\item{lag_dof}{Integer. Degrees of freedom in exposure function for arglag
(see dlnm:crossbasis). Defaults to 4.}

\item{lag_days}{Integer. Maximum lag. Defaults to 3.
(see dlnm:crossbasis).}
}
\value{
A list of cross-basis matrices by region
}
\description{
Creates a cross-basis matrix for each region
}
