% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_pop_totals}
\alias{mh_pop_totals}
\title{Create population totals}
\usage{
mh_pop_totals(df_list, country = "National", meta_analysis = FALSE)
}
\arguments{
\item{df_list}{A list of dataframes containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{country}{Character. Name of country for national level estimates.}

\item{meta_analysis}{Boolean. Whether to perform a meta-analysis.}
}
\value{
List of population totals by year and region
}
\description{
Creates a list of population totals by year and region for use
in the attributable rate calculations.
}
