% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{mh_save_results}
\alias{mh_save_results}
\title{Save results of analysis - Mental Health}
\usage{
mh_save_results(
  rr_results,
  res_attr_tot,
  attr_yr_list,
  attr_mth_list,
  output_folder_path = NULL
)
}
\arguments{
\item{output_folder_path}{Path to folder where results should be saved.
Defaults to NULL.}

\item{results}{Dataframe containing cumulative relative risk and confidence
intervals from analysis.}
}
\description{
Saves a CSV file of cumulative relative risk and
confidence intervals.
}
