% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_pollution.R
\name{plot_air_pollution_forest_by_lag}
\alias{plot_air_pollution_forest_by_lag}
\title{Plot Relative Risk (RR) by lag}
\usage{
plot_air_pollution_forest_by_lag(
  analysis_results,
  max_lag = 14L,
  output_dir = NULL,
  save_plot = FALSE
)
}
\arguments{
\item{analysis_results}{Processed results with RR/AF/AN/AR with lag variables}

\item{max_lag}{Integer. Maximum lag days. Defaults to 14.}

\item{output_dir}{Character. Directory to save plot. Defaults to NULL.}

\item{save_plot}{Logical. Whether to save the plot. Defaults to FALSE.}
}
\value{
ggplot object
}
\description{
Plot Relative Risk (RR) by lag
}
\keyword{internal}
