% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{plot_and_write_relative_risk}
\alias{plot_and_write_relative_risk}
\title{Plot and write results of analysis}
\usage{
plot_and_write_relative_risk(
  df_list,
  argvar,
  bvar,
  blup,
  regions_df,
  mintempregions,
  output_folder_path
)
}
\arguments{
\item{df_list}{A list of dataframes for each region.}

\item{argvar}{An updated list of arguments ($x, $fun, $knots, $degree, $bound) for cross-basis function.}

\item{bvar}{A basis matrix for the predictor vector.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{regions_df}{A dataframe with two columns. Column 1 is abbreviated
region names. Column 2 is full region names.}

\item{mintempregions}{A named numeric vector. Minimum (optimum) mortality temperature per region.}

\item{output_folder_path}{Path to folder for storing outputs.}
}
\value{
A PDF containing a line plot of temperature versus relative risk per region,
and histogram of temperatures per region. A CSV of relative risk per temperature per region.
}
\description{
Plot and write results of analysis
}
\examples{
output_folder_path = 'myfolder/output/'
}
