% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{plot_and_write_relative_risk}
\alias{plot_and_write_relative_risk}
\title{Plot and write results of analysis}
\usage{
plot_and_write_relative_risk(
  df_list,
  blup = NULL,
  regions_df,
  mintempregions,
  save_fig = TRUE,
  save_csv = TRUE,
  dependent_col,
  independent_col,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas,
  output_folder_path
)
}
\arguments{
\item{df_list}{An alphabetically-ordered
list of dataframes for each region.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{regions_df}{A dataframe with two columns.
Column 1 is abbreviated region names.
Column 2 is user-specified region names.}

\item{mintempregions}{A named numeric vector.
Minimum (optimum) mortality temperature per region.}

\item{output_folder_path}{Path to folder for storing outputs.}
}
\value{
A PDF containing a line plot of temperature
versus relative risk per region,
and histogram of temperatures per region.
A CSV of relative risk per temperature per region.
}
\description{
Plot and write results of analysis
}
\examples{
output_folder_path = 'myfolder/output/'
}
