% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{plot_and_write_relative_risk}
\alias{plot_and_write_relative_risk}
\title{Plot and write results of analysis}
\usage{
plot_and_write_relative_risk(
  df_list,
  blup = NULL,
  mintempregions,
  save_fig = TRUE,
  save_csv = TRUE,
  output_folder_path,
  independent_col1,
  independent_col2,
  independent_col3,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas
)
}
\arguments{
\item{df_list}{An alphabetically-ordered
list of dataframes for each region.}

\item{blup}{A list of BLUPs (best linear unbiased predictions).}

\item{mintempregions}{A named numeric vector.
Minimum (optimum) mortality temperature per region.}

\item{save_fig}{Whether to save output figure (Bool)}

\item{save_csv}{Whether to save output CSVs (Bool)}

\item{output_folder_path}{Path to folder for storing outputs.}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}

\item{indepedent_col1_}{column name of first extra independent
variable to include in regression (excluding temperature,
see config file for formula structure). 'None' if none.}

\item{indepedent_col2_}{column name of second independent
variable to include in regression (excluding temperature,
see config file for formula structure). 'None' if none.}

\item{indepedent_col3_}{column name of third independent
variable to include in regression (excluding temperature,
see config file for formula structure). 'None' if none.}
}
\value{
\itemize{

\item A PDF containing a line plot of temperature versus relative risk per
region, and histogram of temperatures per region.
\item A CSV of relative risk per temperature per region.
\item \code{output_df} A dataframe with relative risk estimates and confidence
intervals across the temperature range for each region.
\item \code{tmean_df} A dataframe with daily mean exposure values for each
region.
}
}
\description{
Plot and write results of analysis
}
\examples{
output_folder_path = 'myfolder/output/'
}
