% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{plot_and_write_relative_risk_all}
\alias{plot_and_write_relative_risk_all}
\title{Plot and write results of analysis}
\usage{
plot_and_write_relative_risk_all(
  df_list,
  cb,
  model,
  mintempregions,
  save_fig = TRUE,
  save_csv = TRUE,
  csv_output_path = NULL,
  dependent_col,
  varfun,
  varper,
  vardegree
)
}
\arguments{
\item{df_list}{An alphabetically-ordered
list of dataframes for each region.}

\item{cb}{The generated crossbasis.}

\item{model}{The generated model.}

\item{mintempregions}{A named numeric vector.
Minimum (optimum) mortality temperature per region.}

\item{save_fig}{Whether to save output figure (Bool)}

\item{save_csv}{Whether to save output CSVs (Bool)}

\item{csv_output_path}{The path to save the csv to, including file name and
file extensions}

\item{dependent_col}{the column name of the
dependent variable of interest e.g. deaths}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}
}
\value{
\itemize{

\item A PDF containing a line plot of temperature versus relative risk per
region, and histogram of temperatures per region.
\item A CSV of relative risk per temperature per region.
\item \code{output_df} A dataframe with relative risk estimates and confidence
intervals across the temperature range for each region.
\item \code{temp_df} A dataframe with daily mean exposure values for each
region.
}
}
\description{
Plot and write results of analysis
}
\examples{
csv_output_path = "directory/sub_directory/file_name.csv"

}
\keyword{internal}
