% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{plot_and_write_relative_risk_all}
\alias{plot_and_write_relative_risk_all}
\title{Plot and write results of analysis}
\usage{
plot_and_write_relative_risk_all(
  df_list,
  mintempregions,
  save_fig = TRUE,
  save_csv = TRUE,
  dependent_col,
  varfun,
  varper,
  vardegree,
  coef,
  vcov,
  output_folder_path
)
}
\arguments{
\item{df_list}{An alphabetically-ordered
list of dataframes for each region.}

\item{mintempregions}{A named numeric vector.
Minimum (optimum) mortality temperature per region.}

\item{save_fig}{Whether to save output figure (Bool)}

\item{save_csv}{Whether to save output CSVs (Bool)}

\item{dependent_col}{the column name of the
dependent variable of interest e.g. deaths}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{coef}{A matrix of coefficients for reduced model.}

\item{vcov}{A list. Co-variance matrices for each region for reduced model.}

\item{output_folder_path}{Path to folder for storing outputs.}
}
\value{
\itemize{

\item A PDF containing a line plot of temperature versus relative risk per
region, and histogram of temperatures per region.
\item A CSV of relative risk per temperature per region.
\item \code{output_df} A dataframe with relative risk estimates and confidence
intervals across the temperature range for each region.
\item \code{tmean_df} A dataframe with daily mean exposure values for each
region.
}
}
\description{
Plot and write results of analysis
}
\examples{
output_folder_path = 'myfolder/output/'
}
