% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot histograms of column distributions.}
\usage{
plot_distributions(
  df,
  columns,
  title,
  xlabs = NULL,
  save_hists = FALSE,
  output_path = NULL
)
}
\arguments{
\item{df}{The dataframe containing the data.}

\item{columns}{The columns to plot distributions for.}

\item{title}{The title of your plot.}

\item{xlabs}{A character vector of x-axis labels (e.g., with units)
corresponding to the columns.}

\item{save_hists}{Whether to save the histograms to file.}

\item{output_path}{The path to save your distributions to.}
}
\description{
Plot histograms of column distributions.
}
