% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarrhea.R
\name{plot_monthly_random_effects}
\alias{plot_monthly_random_effects}
\title{Visualise monthly random effects for selected INLA model}
\usage{
plot_monthly_random_effects(
  combined_data,
  model,
  save_fig = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{combined_data}{Data list from combine_health_climate_data() function.}

\item{model}{The fitted model object.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{output_dir}{Character. The path to save the visualisation to. Defaults to NULL.}
}
\value{
THe monthly random effects plot.
}
\description{
Generates and saves a plot of monthly random effects for different
regions, visualizing their contribution to Diarrhea Incidence Rate.
}
