% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_moving_average}
\alias{plot_moving_average}
\title{Plot the moving average of a column.}
\usage{
plot_moving_average(
  df,
  time_col,
  value_col,
  ma_days,
  ma_sides,
  title,
  save_plot = FALSE,
  output_path = ""
)
}
\arguments{
\item{df}{The dataframe containing the raw data.}

\item{time_col}{The column name of the column containing the timeseries.}

\item{value_col}{The column name of the column containing the value.}

\item{ma_days}{The number of days to use for MA calculations.}

\item{ma_sides}{The number of sides to use for MA calculations (1 or 2).}

\item{title}{The title for your plot.}

\item{save_plot}{Whether or not to save the plot.}

\item{output_path}{The path to output the plot to.}
}
\description{
Plot the moving average of a column.
}
