% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{plot_relative_risk}
\alias{plot_relative_risk}
\title{Read in Relative Risk plot at country, Region, and District level}
\usage{
plot_relative_risk(
  data,
  model,
  param_term,
  level = "country",
  filter_year = NULL,
  output_dir = NULL,
  save_csv = FALSE,
  save_fig = FALSE
)
}
\arguments{
\item{data}{Data list from combine_health_climate_data() function.}

\item{model}{The fitted model from run_inla_models() function.}

\item{param_term}{A character vector or list containing parameter terms such
as "tmax" (maximum temperature) and "rainfall" (precipitation).
Default to "tmax".}

\item{level}{A character vector specifying the spatial disaggregation level.
Can take one of the following values: "country", "region", or "district".
Default to "country".}

\item{filter_year}{Integer. The year to filter to data to. Defaults to NULL.}

\item{output_dir}{Character. The path where the PDF file will be saved. Default to NULL.}

\item{save_csv}{Boolean. If TRUE, saves the RR data to the specified directory.
Defaults to FALSE.}

\item{save_fig}{Boolean. If TRUE, saves the plot to the specified directory.
Defaults to FALSE.}
}
\value{
Relative risk plot at country, region, and district levels.
}
\description{
Plots the relative risk of diarrhea cases by the maximum
temperature and cumulative rainfall at country, Region and District level
}
