% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{plot_results}
\alias{plot_results}
\title{Plot results
Compute the attributable deaths for each city,
with emprical CI estimated using the re-centred bases}
\usage{
plot_results(dlist, argvar, bvar, blup, cities, mintempcity, output_csv_path)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A number.
}
\description{
Plot results
Compute the attributable deaths for each city,
with emprical CI estimated using the re-centred bases
}
