% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{plot_results}
\alias{plot_results}
\title{Plot results of analysis}
\usage{
plot_results(
  df_list,
  argvar,
  bvar,
  blup,
  regions_df,
  mintempregions,
  output_folder_path
)
}
\arguments{
\item{output_folder_path}{}
}
\value{
a plot
}
\description{
Plot results of analysis
}
\examples{
output_folder_path = 'myfolder/output/'
}
