% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{plot_results}
\alias{plot_results}
\title{Plot results of analysis}
\usage{
plot_results(data, cb_list, model_list, save_fig, output_folder_path)
}
\arguments{
\item{data}{Dataframe containing daily timeseries data for a health outcome
and climate variables which may be disaggregated by a particular region.}

\item{cb_list}{List of cross_basis matrices from create_crossbasis function.}

\item{model_list}{List of models produced from case-crossover and DLNM
analysis.}

\item{save_fig}{Boolean. Whether to save the plot as an output.}

\item{output_folder_path}{Path to folder where plots should be saved.}
}
\value{
Plots of cumulative lag exposure-response function for each region
}
\description{
Plots cumulative lag exposure-response function for each region
}
