% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarrhea.R
\name{plot_rr_map}
\alias{plot_rr_map}
\title{Plot Relative Risk Map at District or Region Level}
\usage{
plot_rr_map(
  combined_data,
  model,
  param_term = "tmax",
  level = "District",
  filter_year = NULL,
  output_dir = NULL,
  save_fig = FALSE
)
}
\arguments{
\item{combined_data}{A list returned from the \code{combine_health_climate_data()}
function. This list should include both the health-climate data and the map data.}

\item{model}{The fitted model object returned from the \code{run_inla_models()} function.}

\item{param_term}{A character vector or list specifying the climate parameters
(e.g., "tmax" for maximum temperature, "rainfall" for precipitation) to include in the map.
Defaults to "tmax".}

\item{level}{A character string indicating the spatial aggregation level.
Options are "region" or "district". Defaults to "District".}

\item{filter_year}{Integer. The year to filter to data to. Defaults to NULL.}

\item{output_dir}{Character. The directory path where the output PDF file
should be saved. Defaults to NULL.}

\item{save_fig}{Boolean. If TRUE, saves the plot to the specified directory.
Defaults to FALSE.}
}
\value{
Relative risk map at the chosen level.
}
\description{
Generates a map of the relative risk of diarrhea cases associated with climate
hazards, including extreme temperature and cumulative rainfall, at a specified
spatial level (District or Region).
}
