% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{plot_rr_map}
\alias{plot_rr_map}
\title{Plot Relative Risk Map at sub-national Level}
\usage{
plot_rr_map(
  combined_data,
  model,
  param_term = "tmax",
  max_lag,
  nk,
  level = "district",
  case_type,
  filter_year = NULL,
  output_dir = NULL,
  save_fig = FALSE,
  save_csv = FALSE,
  cumulative = FALSE
)
}
\arguments{
\item{combined_data}{A list returned from the \code{combine_health_climate_data()}
function. This list should include both the health-climate data and the map data.}

\item{model}{The fitted model object returned from the \code{run_inla_models()} function.}

\item{param_term}{A character vector or list specifying the climate parameters
(e.g., \code{tmax} for maximum temperature, \code{rainfall} for precipitation) to
include in the map. Defaults to \code{tmax}.}

\item{level}{A character string indicating the spatial aggregation level.
Options are \code{region} or \code{district}. Defaults to \code{district}.}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of \code{diarrhea} or \code{malaria}.}

\item{filter_year}{Integer. The year to filter to data to. Defaults to NULL.}

\item{output_dir}{Character. The directory path where the output PDF file
should be saved. Defaults to NULL.}

\item{save_fig}{Boolean. If TRUE, saves the plot to the specified directory.
Defaults to FALSE.}

\item{cumulative}{Boolean. If TRUE, plot and save cumulative risk of all year
for the specific exposure at region and district level. Defaults to FALSE.}
}
\value{
Relative risk map at the chosen level.
}
\description{
Generates a map of the relative risk of the diseases cases associated with climate
hazards, including extreme temperature and cumulative rainfall, at a specified
geographical level (district or region).
}
\keyword{internal}
