% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{read_and_format_data}
\alias{read_and_format_data}
\title{Code for producing analysis for health effects of extreme weather events - wildfires
Read in and format health data}
\usage{
read_and_format_data(
  health_path,
  date_col,
  mean_temperature_col,
  health_outcome_col,
  region_col = NULL,
  rh_col = NULL,
  wind_speed_col = NULL,
  population_col = NULL
)
}
\arguments{
\item{health_path}{Path to a CSV file containing a daily time series of data
for a particular health outcome and climate variables, which may be disaggregated by region.}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date. Date column should be in YYYY-MM-DD or YYYY/MM/DD format.}

\item{mean_temperature_col}{Character. Name of the column in the dataframe
that contains the daily mean temperature column.}

\item{health_outcome_col}{Character. Name of the column in the dataframe that
contains the daily health outcome count (e.g. number of deaths, hospital
admissions)}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names. Defaults to NULL.}

\item{rh_col}{Character. Name of the column in the dataframe that contains daily relative humidity
values. Defaults to NULL.}

\item{wind_speed_col}{Character. Name of the column in the dataframe that contains daily wind speed.
Defaults to NULL.}

\item{population_col}{Character. Name of the column in the dataframe that
contains the population data.}
}
\value{
Dataframe with formatted and renamed columns
}
\description{
Reads in a CSV file for a daily time series of health and climate data,
renames columns to standard names. Creates columns for day of week, month,
and year columns derived from the date.
}
\keyword{internal}
