% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{run_inla_models}
\alias{run_inla_models}
\title{Run models of increasing complexity in INLA: Fit a baseline model including
spatiotemporal random effects.}
\usage{
run_inla_models(
  combined_data,
  basis_matrices_choices,
  inla_param,
  max_lag,
  nk,
  case_type,
  output_dir = NULL,
  save_model = FALSE,
  family = "nbinomial",
  config = FALSE
)
}
\arguments{
\item{combined_data}{A dataframe resulting from \code{combine_health_climate_data()} function.}

\item{basis_matrices_choices}{A character vector specifying the basis matrix
parameters to be included in the model. Possible values are \code{tmax}, and \code{rainfall}.}

\item{inla_param}{A character vector specifying the confounding exposures to
be included in the model. Possible values are \code{tmax},\code{tmin}, \code{rainfall},
\code{r_humidity}, \code{runoff}, \code{ndvi}, Etc.}

\item{case_type}{Character. The type of disease that the case column refers
to. Must be one of \code{diarrhea} or \code{malaria}.}

\item{output_dir}{Character. The path to save model output to.  Defaults to NULL.}

\item{save_model}{Boolean. Whether to save the results as a CSV. Defaults to
FALSE.}

\item{family}{Character. The probability distribution for the response
variable. The user may also have the possibility to choose \code{nbinomial} for a
negative binomial distribution. Defaults to\code{poisson}.}

\item{config}{Boolean. Enable additional model configurations. Defaults to FALSE.}
}
\value{
A list containing the model, baseline_model, and the dic_table.
}
\description{
: Create and run multiple INLA (Integrated Nested
Laplace Approximation) models to the dataset, evaluates them using
DIC (Deviance Information Criterion), and identifies the best-fitting model.
}
\keyword{internal}
