% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{run_meta_model}
\alias{run_meta_model}
\title{Meta-analysis model}
\usage{
run_meta_model(df_list, regions_df, coef, vcov)
}
\arguments{
\item{df_list}{A list of dataframes for each region}

\item{regions_df}{A dataframe with two columns. Column 1 is abbreviated
region names. Column 2 is full region names.}

\item{coef}{Matrix of coefficients for reduced model}

\item{vcov}{A list. Co-variance matrix for reduced model.}
}
\value{
\itemize{
\item mvmeta model (multivariante meta-analysis)
\item blup BLUP (best linear unbiased predictions) for an mvmeta model
}
\description{
Runs meta-analysis model and estimates best linear unbiased predictions
(BLUPs) from this model.
}
