% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{run_meta_model}
\alias{run_meta_model}
\title{Meta-analysis model}
\usage{
run_meta_model(df_list, regions_df, coef, vcov)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{regions_df}{A dataframe with two columns. Column 1 is abbreviated
region names. Column 2 is user-specified region names.}
}
\value{
\itemize{
\item \code{mvmeta} A model object. A multivariate meta-analysis model.
\item \code{blup} A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each region.
}
}
\description{
Runs meta-analysis model and estimates best linear unbiased predictions
(BLUPs) from this model.
}
