% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{run_meta_model}
\alias{run_meta_model}
\title{Meta-analysis model}
\usage{
run_meta_model(df_list, coef, vcov)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{coef}{A matrix of coefficients for reduced model.}

\item{vcov}{A list. Co-variance matrices for each region for reduced model.}
}
\value{
\itemize{
\item \code{mvmeta} A model object. A multivariate meta-analysis model.
\item \code{blup} A list. BLUP (best linear unbiased predictions) from the
meta-analysis model for each region.
}
}
\description{
Runs meta-analysis model and estimates best linear unbiased predictions
(BLUPs) from this model.
}
