% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasparrini_functions.R
\name{run_model}
\alias{run_model}
\title{Define and run regression model for each dataframe}
\usage{
run_model(df_list, regions_df)
}
\arguments{
\item{regions_df}{A dataframe with two columns. Column 1 is abbreviated
region names. Column 2 is full region names.}

\item{df_list_unordered}{list of dataframes for each region}
}
\value{
\itemize{
\item argvar arguments ($fun, $knots, $degree) for cross-basis function
\item coef matrix of coefficients for reduced model
\item vcov co-variance matrix for reduced model
}
\description{
Define and run regression model for each dataframe
}
