% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{run_model}
\alias{run_model}
\title{Define and run poisson regression model for each dataframe}
\usage{
run_model(df_list, regions_df)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{regions_df}{A dataframe with two columns. Column 1 is abbreviated
region names. Column 2 is user-specified region names.}
}
\value{
\itemize{
\item \code{coef} A matrix of coefficients for reduced model.
\item \code{vcov} A list. Co-variance matrices for each region for reduced model.
}
}
\description{
Define and run poisson regression model for each dataframe
}
