% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{run_model}
\alias{run_model}
\title{Define and run poisson regression model for each dataframe}
\usage{
run_model(
  df_list,
  independent_cols = NULL,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas
)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{independent_cols}{column name (or list of names) of extra independent
variable to include in regression (excluding temperature). Defaults to NULL.}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}
}
\value{
\itemize{
\item \code{coef_} A matrix of coefficients for reduced model.
\item \code{vcov_} A list. Co-variance matrices for each region for reduced model.
\item \code{cb} Basis matrices for the two dimensions of predictor and lags.
\item \code{model} A quasi-poission generalised linear model object.
See: https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/glm
}
}
\description{
Define and run poisson regression model for each dataframe
}
