% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_dlnm_funcs.R
\name{run_model}
\alias{run_model}
\title{Define and run poisson regression model for each dataframe}
\usage{
run_model(
  df_list,
  independent_col1,
  independent_col2,
  independent_col3,
  varfun,
  varper,
  vardegree,
  lag,
  lagnk,
  dfseas
)
}
\arguments{
\item{df_list}{An alphabetically-ordered list of dataframes for each region.}

\item{varfun}{Exposure function
(see dlnm::crossbasis)}

\item{varper}{Internal knot positions in exposure function
(see dlnm::crossbasis)}

\item{vardegree}{Degrees of freedom in exposure function
(see dlnm:crossbasis)}

\item{lag}{Lag length in time
(see dlnm::logknots)}

\item{lagnk}{Number of knots in lag function
(see dlnm::logknots)}

\item{dfseas}{Degrees of freedom for seasonality}

\item{indepedent_col1_}{column name of first extra independent
variable to include in regression (excluding temperature,
see config file for formula structure). 'None' if none.}

\item{indepedent_col2_}{column name of second independent
variable to include in regression (excluding temperature,
see config file for formula structure). 'None' if none.}

\item{indepedent_col3_}{column name of third independent
variable to include in regression (excluding temperature,
see config file for formula structure). 'None' if none.}
}
\value{
\itemize{
\item \code{coef} A matrix of coefficients for reduced model.
\item \code{vcov} A list. Co-variance matrices for each region for reduced model.
}
}
\description{
Define and run poisson regression model for each dataframe
}
