% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{save_results}
\alias{save_results}
\title{Save results of analysis}
\usage{
save_results(
  rr_results,
  output_AN_AR = TRUE,
  an_ar_results = NULL,
  output_folder_path
)
}
\arguments{
\item{rr_results}{Dataframe. Relative risk and confidence intervals for
each lag of wildfire-related PM2.5}

\item{output_AN_AR}{Bool. Whether or not to output attributable numbers and
fractions. Defaults to TRUE.}

\item{an_ar_results}{Dataframe. The AN/AR results. Defaults to NULL.}

\item{output_folder_path}{Path to folder where results should be saved.}
}
\description{
Saves a CSV file of relative risk and confidence intervals for
each lag value of wildfire-related PM2.5
}
\keyword{internal}
