% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{save_results}
\alias{save_results}
\title{Save results of analysis}
\usage{
save_results(results, output_folder_path)
}
\arguments{
\item{results}{Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5}

\item{output_folder_path}{Path to folder where results should be saved.}
}
\description{
Saves a CSV file of relative risk and confidence intervals for
each lag value of wildfire-related PM2.5
}
