% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{save_wildfire_results}
\alias{save_wildfire_results}
\title{Save results of wildfire related analysis}
\usage{
save_wildfire_results(rr_results, an_ar_results = NULL, output_folder_path)
}
\arguments{
\item{an_ar_results}{Dataframe containing attributable number/fraction
results. Defaults to NULL.}

\item{output_folder_path}{Path to folder where results should be saved.}

\item{results}{Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5.}
}
\description{
Saves a CSV file of relative risk and confidence intervals for
each lag value of wildfire-related PM2.5. Also optionally save results of
attributable numbers/fractions.
}
\keyword{internal}
