% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{set_cross_basis}
\alias{set_cross_basis}
\title{Create cross-basis matrices for DLNM analysis}
\usage{
set_cross_basis(data, include_cvh = FALSE)
}
\arguments{
\item{data}{A dataset returned from \code{combine_health_climate_data()},
including lagged variables like \code{tmax_lag1}, \code{tmin_lag1}, etc.}

\item{include_cvh}{Logical. Whether or not to include the cvh column in
the var definitions. Needed for Malaria.}
}
\value{
A named list of cross-basis matrices for available climate variables:
maximum temperature, minimum temperature, rainfall, relative humidity, etc.
}
\description{
Generates cross-basis matrices for lagged climate variables in a dataset,
for use in Distributed Lag Nonlinear Models (DLNM).
}
