% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases_shared.R
\name{set_cross_basis}
\alias{set_cross_basis}
\title{Create a cross-basis matrix set for DLNM analysis}
\usage{
set_cross_basis(data, max_lag, include_ndvi = FALSE)
}
\arguments{
\item{data}{A dataset returned from \code{combine_health_climate_data()},
including lagged variables like \code{tmax_lag1}, \code{tmin_lag1}, etc.}

\item{max_lag}{Character. Number corresponding to the maximum lag to be
considered for the delay effect. It should be between 2 an 4. Defaults to 2.}
}
\value{
A list of cross-basis matrices including the basis matrix for maximum
temperature, minimun temperature, cumulative rainfall, and relative humidity.
}
\description{
Generates cross-basis matrices for lagged climate variables in a dataset,
for use in Distributed Lag Nonlinear Models (DLNM).
}
\keyword{internal}
