% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarrhea.R
\name{set_cross_basis}
\alias{set_cross_basis}
\title{Create a cross-basis matrix set for DLNM analysis}
\usage{
set_cross_basis(data)
}
\arguments{
\item{data}{The dataset containing district_code, region_code and year
columns from the combine_health_climate_data() function.}
}
\value{
A list of cross-basis matrices including the basis matrix for maximum
temperature, minimun temperature, cumulative rainfall, and relative humidity.
}
\description{
Creates cross-basis matrix for each climate variable.
}
