% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{suicides_heat_do_analysis}
\alias{suicides_heat_do_analysis}
\title{Run pipeline to analyse the impact of extreme heat on suicides using a time-
stratified case-crossover approach with distributed lag non-linear model}
\usage{
suicides_heat_do_analysis(
  data_path,
  date_col,
  region_col = NULL,
  temperature_col,
  health_outcome_col,
  population_col,
  var_fun = "bs",
  var_degree = 2,
  var_per = c(25, 50, 75),
  lag_fun = "strata",
  lag_breaks = 1,
  lag_days = 2,
  save_fig = FALSE,
  save_csv = FALSE,
  cenper = 50,
  country = "National",
  descriptive_stats = FALSE,
  ds_correlation_method = "pearson",
  ds_use_individual_dfs = TRUE,
  ds_dist_columns = c(),
  ds_ma_days = 100,
  ds_ma_sides = 2,
  ds_ma_columns = c(),
  output_folder_path = NULL
)
}
\arguments{
\item{data_path}{Path to a csv file containing a daily time series of data
for a particular health outcome and climate variables, which may be
disaggregated by region.}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names. Defaults to NULL.}

\item{temperature_col}{Character. Name of the column in the dataframe that
contains the temperature column.}

\item{health_outcome_col}{Character. Name of the column in the dataframe that
contains the health outcome count column (e.g. number of deaths, hospital
admissions).}

\item{var_fun}{Character. Exposure function for argvar
(see dlnm::crossbasis). Defaults to 'bs'.}

\item{var_degree}{Integer. Degree of the piecewise polynomial for argvar
(see dlnm:crossbasis). Defaults to 2 (quadratic).}

\item{var_per}{Vector. Internal knot positions for argvar
(see dlnm::crossbasis). Defaults to c(25,50,75).}

\item{lag_fun}{Character. Exposure function for arglag
(see dlnm::crossbasis). Defaults to 'strata'.}

\item{lag_breaks}{Integer. Internal cut-off point defining the strata for arglag
(see dlnm:crossbasis). Defaults to 1.}

\item{lag_days}{Integer. Maximum lag. Defaults to 2.
(see dlnm:crossbasis).}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{save_csv}{Boolean. Whether to save the results as a CSV. Defaults to
FALSE.}

\item{cenper}{Integer. Value for the percentile in calculating the centering
value 0-100. Defaults to 50.}

\item{country}{Character. Name of country for national level estimates.}

\item{descriptive_stats}{Boolean. Whether to calculate descriptive stats.}

\item{ds_correlation_method}{character. The correlation method used in correlation matrices.
Defaults to 'pearson'.}

\item{ds_use_individual_dfs}{Boolean. Whether to calculate descriptive stats for each individual
df in df_list. Default to TRUE.}

\item{ds_dist_columns}{character vector. The names of columns to plot distributions for.
Defaults to c().}

\item{ds_ma_days}{integer. How many days to use for moving average calculations.
Defaults to 100.}

\item{ds_ma_sides}{integer. How many sides to use for moving average calculations (1 or 2).
Defaults to 2.}

\item{ds_ma_columns}{character vector. The names of columns to plot moving average for.}

\item{output_folder_path}{Path to folder where plots and/or CSV should be
saved. Defaults to NULL.}
}
\value{
Dataframe containing cumulative relative risk and confidence
intervals from analysis.
}
\description{
Runs full analysis pipeline for analysis of the impact of
extreme heat on suicides
}
