% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health.R
\name{suicides_heat_do_analysis}
\alias{suicides_heat_do_analysis}
\title{Run pipeline to analyse the impact of extreme heat on suicides using a time-
stratified case-crossover approach with distributed lag non-linear model}
\usage{
suicides_heat_do_analysis(
  data_path,
  date_col,
  region_col = NULL,
  temperature_col,
  health_outcome_col,
  var_fun = "ns",
  var_dof = 4,
  lag_fun = "ns",
  lag_dof = 4,
  lag_days = 3,
  save_fig = FALSE,
  save_csv = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{data_path}{Path to a csv file containing a daily time series of data
for a particular health outcome and climate variables, which may be
disaggregated by region.}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names. Defaults to NULL.}

\item{temperature_col}{Character. Name of the column in the dataframe that
contains the temperature column.}

\item{health_outcome_col}{Character. Name of the column in the dataframe that
contains the health outcome count column (e.g. number of deaths, hospital
admissions).}

\item{var_fun}{Character. Exposure function (see dlnm::crossbasis). Defaults
to 'ns'.}

\item{var_dof}{Integer. Degrees of freedom in exposure function
(see dlnm:crossbasis). Defaults to 4.}

\item{lag_fun}{Character. Exposure function (see dlnm::crossbasis). Defaults
to 'ns'.}

\item{lag_dof}{Integer. Degrees of freedom in exposure function
(see dlnm:crossbasis). Defaults to 4.}

\item{lag_days}{Integer. Maximum lag (see dlnm:crossbasis). Defaults to 3.}

\item{save_fig}{Boolean. Whether to save the plot as an output. Defaults to
FALSE.}

\item{save_csv}{Boolean. Whether to save the results as a CSV. Defaults to
FALSE.}

\item{output_folder_path}{Path to folder where plots and/or CSV should be
saved. Defaults to NULL.}

\item{descriptive_stats}{Boolean. Whether to calculate descriptive stats.}

\item{ds_correlation_method}{character. The correlation method used in correlation matrices.
Defaults to 'pearson'.}

\item{ds_use_individual_dfs}{Boolean. Whether to calculate descriptive stats for each individual
df in df_list. Default to TRUE.}

\item{ds_dist_columns}{character vector. The names of columns to plot distributions for.
Defaults to c().}

\item{ds_ma_days}{integer. How many days to use for moving average calculations.
Defaults to 100.}

\item{ds_ma_sides}{integer. How many sides to use for moving average calculations (1 or 2).
Defaults to 2.}

\item{ds_ma_columns}{character vector. The names of columns to plot moving average for.}
}
\value{
Dataframe containing cumulative relative risk and confidence
intervals from analysis.
}
\description{
Runs full analysis pipeline for analysis of the impact of
extreme heat on suicides
}
