% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{summarise_AF_AN}
\alias{summarise_AF_AN}
\title{Summarise AF and AN numbers by region and year}
\usage{
summarise_AF_AN(data)
}
\arguments{
\item{data}{Dataframe containing daily data including calculated AF and AN.}
}
\value{
Dataframe containing summarised AF and AN data, by year and region
}
\description{
Takes daily data with attributable fraction and attributable number
and summarises by year and region
}
