% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{wildfire_do_analysis}
\alias{wildfire_do_analysis}
\title{Run pipeline to analyse the impact of wildfire-related PM2.5 on a health
outcome using a time-stratified case-crossover approach.}
\usage{
wildfire_do_analysis(
  health_path,
  join_wildfire_data = FALSE,
  ncdf_path = NULL,
  shp_path = NULL,
  date_col,
  region_col,
  mean_temperature_col,
  health_outcome_col,
  wildfire_lag = 3,
  temperature_lag = 1,
  spline_temperature_lag = 0,
  spline_temperature_degrees_freedom = 6,
  predictors_vif = NULL,
  relative_risk_by_region = FALSE,
  scale_factor_wildfire_pm = 10,
  save_fig = FALSE,
  save_csv = FALSE,
  output_folder_path = NULL,
  print_vif = FALSE,
  print_model_summaries = FALSE
)
}
\arguments{
\item{health_path}{Path to a CSV file containing a daily time series of data
for a particular health outcome, which may be disaggregated by region. If
this does not include a column with wildfire-related PM2.5, use
join_wildfire_data = TRUE to join these data.}

\item{join_wildfire_data}{Boolean. If TRUE, a daily time series of
wildfire-related PM2.5 concentration is joined to the health data. If FALSE,
the data set is loaded without any additional joins.}

\item{ncdf_path}{Path to a NetCDF file containing a daily time series of
gridded wildfire-related PM2.5 concentration data.}

\item{shp_path}{Path to a shapefile .shp of the geographical boundaries for
which to extract mean values of wildfire-related PM2.5}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names.}

\item{mean_temperature_col}{Character. Name of the column in the dataframe
that contains the mean temperature column.}

\item{health_outcome_col}{Character. Name of the column in the dataframe that
contains the health outcome count column (e.g. number of deaths, hospital
admissions)}

\item{wildfire_lag}{Integer. The number of days for which to calculate the
lags for wildfire PM2.5. Default is 3.}

\item{temperature_lag}{Integer. The number of days for which to calculate
the lags for temperature. Default is 1.}

\item{spline_temperature_lag}{Integer. The number of days of lag in the
temperature variable from which to generate splines. Default is 0 (unlagged
temperature variable).}

\item{spline_temperature_degrees_freedom}{Integer. Degrees of freedom for the
spline(s).}

\item{predictors_vif}{Character vector with each of the predictors to
include in the model. Must contain at least 2 variables. Defaults to NULL.}

\item{relative_risk_by_region}{Bool. Whether to calculate Relative Risk by region.
Default: FALSE}

\item{scale_factor_wildfire_pm}{Numeric. The value to divide the wildfire
PM2.5 concentration variables by for alternative interpretation of outputs.
Corresponds to the unit increase in wildfire PM2.5 to give the model
estimates and relative risks (e.g. scale_factor = 10 corresponds to estimates
and relative risks representing impacts of a 10 unit increase in wildfire
PM2.5). Setting this parameter to 0 or 1 leaves the variable unscaled.}

\item{save_fig}{Boolean. Whether to save the plot as an output.}

\item{save_csv}{Boolean. Whether to save the results as a CSV}

\item{output_folder_path}{Path. Path to folder where plots and/or CSV should
be saved.}

\item{print_vif}{Bool, whether or not to print VIF (variance inflation factor)
for each predictor. Defaults to FALSE.}

\item{print_model_summaries}{Bool. Whether to print the model summaries to
console. Defaults to FALSE.}
}
\value{
Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5
}
\description{
Runs full analysis pipeline for analysis of the impact of
wildfire-related PM2.5 on a health outcome.
}
