% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{wildfire_do_analysis}
\alias{wildfire_do_analysis}
\title{Run analysis pipeline for indicator of the health impacts of wildfire-related
PM2.5.}
\usage{
wildfire_do_analysis(
  health_path,
  join_wildfire_data,
  ncdf_path,
  shp_path,
  date_col,
  region_col = region_col,
  mean_temperature_col,
  health_outcome_col,
  wildfire_lag = 3,
  temperature_lag = 1,
  spline_temperature_col,
  spline_temperature_df = 6,
  variables_descriptive_stats,
  bin_width_histogram = 10,
  predictors_vif,
  scale_wildfire_pm = TRUE,
  scale_factor_wildfire_pm = 10,
  save_fig = TRUE,
  save_csv = TRUE,
  output_folder_path
)
}
\arguments{
\item{health_path}{Path to a CSV file containing a daily time series of data
for a particular health outcome, which may be disaggregated by region. If
this does not include a column with wildfire-related PM2.5, use
join_wildfire_data = TRUE to join these data.}

\item{join_wildfire_data}{Boolean. If TRUE, a daily time series of
wildfire-related PM2.5 concentration is joined to the health data. If FALSE,
the data set is loaded without any additional joins.}

\item{ncdf_path}{Path to a NetCDF file containing a daily time series of
gridded wildfire-related PM2.5 concentration data.}

\item{shp_path}{Path to a shapefile .shp of the geographical boundaries for
which to extract mean values of wildfire-related PM2.5}

\item{date_col}{Character. Name of the column in the dataframe that contains
the date.}

\item{region_col}{Character. Name of the column in the dataframe that contains
the region names.}

\item{mean_temperature_col}{Character. Name of the column in the dataframe
that contains the mean temperature column.}

\item{health_outcome_col}{Character. Name of the column in the dataframe that
contains the health outcome count column (e.g. number of deaths, hospital
admissions)}

\item{wildfire_lag}{Integer. The number of days for which to calculate the
lags for wildfire PM2.5. Default is 3.}

\item{temperature_lag}{Integer. The number of days for which to calculate
the lags for temperature. Default is 1.}

\item{spline_temperature_col}{Character. The name of the temperature column
in the dataframe from which to generate splines. This could be lagged or
non-lagged temperature.}

\item{spline_temperature_df}{Integer. Degrees of freedom for the spline(s).}

\item{variables_descriptive_stats}{Character or character vector with
variable to produce summary statistics for. Must include at least 1 variable.}

\item{bin_width_histogram}{Integer. Width of each bin in a histogram of the
outcome variable. Default is 5.}

\item{predictors_vif}{Character vector with each of the predictors to
include in the model. Must contain at least 2 variables.}

\item{scale_wildfire_pm}{Boolean. Whether to scale wildfire PM2.5
concentration variable for alternative interpretation of outputs. Default is
TRUE.}

\item{scale_factor_wildfire_pm}{Numeric. If scale_wildfire_pm = TRUE, the
value to divide the wildfire PM2.5 concentration variables by. Corresponds to
the unit increase in wildfire PM2.5 to give the model estimates and relative
risks (e.g. scale_factor = 10 corresponds to estimates and relative risks
representing impacts of a 10 unit increase in wildfire PM2.5)}

\item{save_fig}{Boolean. Whether to save the plot as an output.}

\item{save_csv}{Boolean. Whether to save the results as a CSV}

\item{output_folder_path}{Path to folder where plots and/or CSV should be
saved.}
}
\value{
Dataframe of relative risk and confidence intervals for
each lag of wildfire-related PM2.5
}
\description{
Runs full analysis pipeline for analysis of impacts of
wildfire-related PM2.5 on health outcome.
}
