% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{write_attributable_deaths}
\alias{write_attributable_deaths}
\title{Write outputs to csv}
\usage{
write_attributable_deaths(
  avgtmean_wald,
  rangetmean_wald,
  anregions_bind,
  antot_bind,
  arregions_bind,
  artot_bind,
  save_csv = FALSE,
  output_folder_path = NULL
)
}
\arguments{
\item{avgtmean_wald}{The Wald statistic P-value for the average temperature.}

\item{rangetmean_wald}{The wald statistic P-value for the temperature range.}

\item{anregions_bind}{A dataframe with attributable deaths for each region.}

\item{antot_bind}{A matrix of numbers of numbers of deaths attributable to
temperature, heat, cold, extreme heat and extreme cold (with confidence
intervals).}

\item{arregions_bind}{A dataframe with attributable rates by region.}

\item{artot_bind}{A matrix of fractions of all-cause mortality
attributable to temperature, heat, cold, extreme heat and extreme cold
(with confidence intervals).}

\item{save_csv}{Bool. Whether or not to save CSV files to output path.
Defaults to FALSE.}

\item{output_folder_path}{Path to folder for storing outputs.}
}
\value{
\itemize{
\item \code{wald_publication} A dataframe containing the Wald statistic P-values
for average temperature and the temperature range.
\item \code{anregions_publication} A matrix of numbers of deaths attributable to
temperature, heat, cold, extreme heat and extreme cold (with confidence
intervals), disaggregated by region.
\item \code{antot_bind} A matrix of numbers of numbers of deaths attributable to
temperature, heat, cold, extreme heat and extreme cold (with confidence
intervals).
\item \code{arregions_publication}
\item \code{artot_bind} A matrix of fractions of all-cause mortality
attributable to temperature, heat, cold, extreme heat and extreme cold
(with confidence intervals).
}
}
\description{
Write the attributable deaths and temperature for each regions,
with empirical CI estimated using the re-centered bases.
}
\examples{
output_folder_path = 'myfolder/output/'

}
